% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acaps_government_measures_data.R
\name{acaps_government_measures_data}
\alias{acaps_government_measures_data}
\title{ACAPS government measures catalogs worldwide COVID-19 government interventions}
\usage{
acaps_government_measures_data()
}
\value{
a data.frame
}
\description{
ACAPS government measures catalogs worldwide COVID-19 government interventions
}
\details{
The #COVID19 Government Measures Dataset puts together all the measures implemented by governments worldwide in response to the Coronavirus pandemic. Data collection includes secondary data review. The researched information available falls into five categories:
\itemize{
\item Social distancing
\item Movement restrictions
\item Public health measures
\item Social and economic measures
\item Lockdowns
}

Each category is broken down into several types of measures.

ACAPS government measures catalogs worldwide COVID-19 government interventionsACAPS consulted government, media, United Nations, and other organisations sources.
}
\examples{
res = acaps_government_measures_data()
head(res)
colnames(res)
dplyr::glimpse(res)


}
\references{
\itemize{
\item \url{https://www.acaps.org/covid19-government-measures-dataset}
}
}
\seealso{
Other data-import: 
\code{\link{acaps_secondary_impact_data}()},
\code{\link{apple_mobility_data}()},
\code{\link{beoutbreakprepared_data}()},
\code{\link{cdc_aggregated_projections}()},
\code{\link{cdc_excess_deaths}()},
\code{\link{cdc_social_vulnerability_index}()},
\code{\link{coronadatascraper_data}()},
\code{\link{coronanet_government_response_data}()},
\code{\link{cov_glue_lineage_data}()},
\code{\link{cov_glue_newick_data}()},
\code{\link{cov_glue_snp_lineage}()},
\code{\link{covidtracker_data}()},
\code{\link{descartes_mobility_data}()},
\code{\link{ecdc_data}()},
\code{\link{economist_excess_deaths}()},
\code{\link{financial_times_excess_deaths}()},
\code{\link{google_mobility_data}()},
\code{\link{government_policy_timeline}()},
\code{\link{jhu_data}()},
\code{\link{jhu_us_data}()},
\code{\link{kff_icu_beds}()},
\code{\link{nytimes_county_data}()},
\code{\link{oecd_unemployment_data}()},
\code{\link{owid_data}()},
\code{\link{param_estimates_published}()},
\code{\link{test_and_trace_data}()},
\code{\link{us_county_geo_details}()},
\code{\link{us_county_health_rankings}()},
\code{\link{us_healthcare_capacity}()},
\code{\link{us_hospital_details}()},
\code{\link{us_state_distancing_policy}()},
\code{\link{usa_facts_data}()},
\code{\link{who_cases}()},
\code{\link{yougov_behavioral_data}()}

Other NPI: 
\code{\link{coronanet_government_response_data}()},
\code{\link{government_policy_timeline}()},
\code{\link{us_state_distancing_policy}()}
}
\author{
Sean Davis \href{mailto:seandavi@gmail.com}{seandavi@gmail.com}
}
\concept{NPI}
\concept{data-import}
