% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acaps_secondary_impact_data.R
\name{acaps_secondary_impact_data}
\alias{acaps_secondary_impact_data}
\title{ACAPS secondary effects of COVID-19}
\usage{
acaps_secondary_impact_data()
}
\value{
a data.frame
}
\description{
The objective of the dataset is to provide information for decision makers to improve their efforts in addressing the wider effects of the COVID-19 pandemic.
}
\details{
The ACAPS COVID-19 Analytical Framework, together with the data collected for the Government Measures dataset, provided the foundation for the Secondary Impacts Analytical Framework and dataset. The dataset will track secondary impacts across a wide range of relevant themes such as economy, health, migration and education.

Around 80 impact indicators, anticipated to be impacted by COVID-19, have been identified and organised across 4 pillars and 13 thematic blocks.

The data collection is ongoing and will be conducted on a country-level. Eventually, data will identify the secondary impacts of the COVID-19 pandemic in more than 190 countries. Data comes from a range of available sources including international organisations, research centres and media analysis.

The data collection is supported by a group of student volunteers from various universities worldwide. Volunteers receive training on the analytical framework and indicators to ensure coherent data. Additional guidance on analytical and data collection techniques are also provided by an ACAPS analyst team which supervises the data entered. This model has already been successfully implemented in the ACAPS Government Measures project.
}
\examples{

res = acaps_secondary_impact_data()

head(res)
dplyr::glimpse(res)
dim(res)

}
\references{
\itemize{
\item https://www.acaps.org/secondary-impacts-covid-19
}
}
\seealso{
Other data-import: 
\code{\link{acaps_government_measures_data}()},
\code{\link{apple_mobility_data}()},
\code{\link{beoutbreakprepared_data}()},
\code{\link{cci_us_vaccine_data}()},
\code{\link{cdc_aggregated_projections}()},
\code{\link{cdc_excess_deaths}()},
\code{\link{cdc_social_vulnerability_index}()},
\code{\link{coronadatascraper_data}()},
\code{\link{coronanet_government_response_data}()},
\code{\link{cov_glue_lineage_data}()},
\code{\link{cov_glue_newick_data}()},
\code{\link{cov_glue_snp_lineage}()},
\code{\link{covidtracker_data}()},
\code{\link{descartes_mobility_data}()},
\code{\link{ecdc_data}()},
\code{\link{econ_tracker_unemp_data}},
\code{\link{economist_excess_deaths}()},
\code{\link{financial_times_excess_deaths}()},
\code{\link{google_mobility_data}()},
\code{\link{government_policy_timeline}()},
\code{\link{jhu_data}()},
\code{\link{jhu_us_data}()},
\code{\link{kff_icu_beds}()},
\code{\link{nytimes_county_data}()},
\code{\link{oecd_unemployment_data}()},
\code{\link{owid_data}()},
\code{\link{param_estimates_published}()},
\code{\link{test_and_trace_data}()},
\code{\link{us_county_geo_details}()},
\code{\link{us_county_health_rankings}()},
\code{\link{us_healthcare_capacity}()},
\code{\link{us_hospital_details}()},
\code{\link{us_state_distancing_policy}()},
\code{\link{usa_facts_data}()},
\code{\link{who_cases}()}

Other economics: 
\code{\link{econ_tracker_unemp_data}},
\code{\link{us_county_health_rankings}()}

Other social: 
\code{\link{us_county_health_rankings}()}
}
\author{
Sean Davis \href{mailto:seandavi@gmail.com}{seandavi@gmail.com}
}
\concept{data-import}
\concept{economics}
\concept{social}
