% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_catalog.R
\name{add_dataset_to_catalog}
\alias{add_dataset_to_catalog}
\title{add a dataset to the package catalog}
\usage{
add_dataset_to_catalog(
  name,
  accessor,
  data_type = list(),
  url = "",
  geographical = TRUE,
  geospatial = FALSE,
  resolution = NULL,
  region = NULL,
  jsonfile = "inst/data_catalog/catalog.json",
  write = FALSE
)
}
\arguments{
\item{name}{character(1) long name}

\item{accessor}{character(1) name of accessor function}

\item{data_type}{character() vector of data_types}

\item{url}{character() url of source}

\item{geographical}{logical(1) includes geographical information or not}

\item{geospatial}{logical(1) includes geospatial information or not}

\item{resolution}{logical(1) if geographical, one of admin0, admin1, admin2, admin3}

\item{region}{character() with World (global), international (more than one nation),
United States}

\item{jsonfile}{character(1) name of json file to append to.}

\item{write}{logical(1) write the file (TRUE) or simply (invisibly) return the
resulting new list for inspection}
}
\value{
A list ready for appending to json
}
\description{
This is an \emph{internal} convenience function for adding a new dataset to the
package catalog.
}
\author{
Sean Davis \href{mailto:seandavi@gmail.com}{seandavi@gmail.com}
}
