% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apple_mobility_data.R
\name{apple_mobility_data}
\alias{apple_mobility_data}
\title{Access Apple mobility data}
\usage{
apple_mobility_data(agree_to_terms = TRUE, max_tries = 3, message_url = FALSE)
}
\arguments{
\item{agree_to_terms}{logical, when TRUE, implies that the user
has agreed to Apple's terms of use. See references and note.}

\item{max_tries}{integer, the number of tries to attempt downloading}

\item{message_url}{logical, output a message with the URL for the day
since Apple changes it daily.}
}
\description{
From Apple's website:
"Learn about COVID‑19 mobility trends in countries/regions and cities. Reports are published daily and reflect requests for directions in Apple Maps. Privacy is one of our core values, so Maps doesn’t associate your data with your Apple ID, and Apple doesn’t keep a history of where you’ve been."
}
\details{
The CSV file and charts on this site show a relative volume of
directions requests per country/region or city compared to a
baseline volume on January 13th, 2020.

We define our day as midnight-to-midnight, Pacific time. Cities
represent usage in greater metropolitan areas and are stably
defined during this period. In many countries/regions and cities,
relative volume has increased since January 13th, consistent with
normal, seasonal usage of Apple Maps. Day of week effects are
important to normalize as you use this data.

Data that is sent from users’ devices to the Maps service is
associated with random, rotating identifiers so Apple doesn’t have
a profile of your movements and searches. Apple Maps has no
demographic information about our users, so we can’t make any
statements about the representativeness of our usage against the
overall population.

These data are available from a URL that changes daily. The parent
page is the place to check to see what is going on if there are problems.
}
\note{
Apple requires that all users agree to their terms of use.
See \url{https://www.apple.com/covid19/mobility}.
}
\examples{

res = apple_mobility_data()
colnames(res)
head(res)
table(res$transportation_type)

require(ggplot2)

pl = res \%>\%
    dplyr::filter(region \%in\% c('Russia','New York City','Italy')) \%>\%
    ggplot(aes(x=date)) +
        geom_line(aes(y=mobility_index,color=transportation_type)) +
        scale_x_date(date_breaks = '1 week', date_labels='\%b-\%d') +
        facet_grid(rows=vars(region)) +
        ggtitle('Changes in Apple Mobility Index over time')
pl

regs_of_interest = c('Seattle', 'New York City',
                     'Chicago', 'Italy',
                     'Russia', 'UK',
                     'Brazil')
res \%>\%
    dplyr::filter(region \%in\% regs_of_interest) \%>\%
    ggplot(aes(x=date, y=region, fill=mobility_index)) +
        geom_tile() +
        facet_grid(rows=vars(transportation_type)) +
        ggtitle('Changes in Apple Mobility Index over time')

if(require(viridis)) {
res \%>\%
    dplyr::filter(region \%in\% regs_of_interest) \%>\%
    ggplot(aes(x=date, y=region, fill=mobility_index)) +
        geom_tile() +
        facet_grid(rows=vars(transportation_type)) +
        scale_fill_viridis() +
        ggtitle('Changes in Apple Mobility Index over time')
}

if(require(plotly)) {
    ggplotly(pl)
}


}
\references{
\itemize{
\item \url{https://www.apple.com/covid19/mobility}
}
}
\seealso{
Other data-import: 
\code{\link{beoutbreakprepared_data}()},
\code{\link{cdc_aggregated_projections}()},
\code{\link{cdc_excess_deaths}()},
\code{\link{cov_glue_lineage_data}()},
\code{\link{cov_glue_newick_data}()},
\code{\link{covidtracker_data}()},
\code{\link{descartes_mobility_data}()},
\code{\link{ecdc_data}()},
\code{\link{eu_data_cache_data}()},
\code{\link{google_mobility_data}()},
\code{\link{government_policy_timeline}()},
\code{\link{healthdata_projections_data}()},
\code{\link{jhu_data}()},
\code{\link{jhu_us_data}()},
\code{\link{kff_icu_beds}()},
\code{\link{nytimes_county_data}()},
\code{\link{param_estimates_published}()},
\code{\link{us_county_geo_details}()},
\code{\link{us_healthcare_capacity}()},
\code{\link{us_state_distancing_policy}()},
\code{\link{usa_facts_data}()}

Other mobility: 
\code{\link{descartes_mobility_data}()},
\code{\link{google_mobility_data}()}
}
\author{
Sean Davis \href{mailto:seandavi@gmail.com}{seandavi@gmail.com}
}
\concept{data-import}
\concept{mobility}
