% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arima.R
\name{Arima_contig_states}
\alias{Arima_contig_states}
\title{full incidence for contiguous states}
\usage{
Arima_contig_states(
  src,
  state.in = "All contig",
  MAorder = 2,
  trend = "linear",
  basedate = "2020-03-15",
  lookback_days = 29,
  ARorder = 0,
  contig_vec = contig_states_twolet()
)
}
\arguments{
\item{state.in}{character(1) state name}

\item{MAorder}{numeric(1) order of moving average component}

\item{trend}{character(1) if "linear" differencing order for ARIMA=1, else 2}

\item{basedate}{character(1) used by lubridate::as_date to filter away all earlier records}

\item{lookback_days}{numeric(1) only uses this many days from most recent in src}

\item{ARorder}{order of autoregressive component}

\item{contig_vec}{character() vector of tokens for subsetting src}
}
\description{
full incidence for contiguous states
}
\examples{
usd = usa_facts_data()
cont = Arima_contig_states(usd)
cont
plot(cont)
}
