% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arima.R
\name{Arima_drop_state}
\alias{Arima_drop_state}
\title{fit ARIMA model to US data dropping one state}
\usage{
Arima_drop_state(
  src_us,
  src_st,
  state.in = "New York",
  MAorder = 2,
  trend = "linear",
  basedate = "2020-03-15",
  lookback_days = 29,
  ARorder = 0
)
}
\arguments{
\item{src_us}{tibble for national level data like that of enriched_jhu_data()}

\item{src_st}{tibble for state level data like that of nytimes_state_data()}

\item{state.in}{character(1) state name}

\item{MAorder}{numeric(1) order of moving average component}

\item{trend}{character(1) if "linear" differencing order for ARIMA=1, else 2}

\item{basedate}{character(1) used by lubridate::as_date to filter away all earlier records}

\item{lookback_days}{numeric(1) only uses this many days from most recent in src}

\item{ARorder}{order of autoregressive component}
}
\value{
instance of S3 class Arima_sars2pack
}
\description{
fit ARIMA model to US data dropping one state
}
\note{
Apparent discrepancies in counts in vicinity of April 15 between
NYT and JHU for full US incidence lead us to employ the two sources
for the example.  If NYT alone is used, summing to obtain USA
overall, the resulting USA incidence trend is
extravagantly variable.
}
\examples{
ej = enriched_jhu_data()
usa_full = Arima_nation(ej)
plot(usa_full)
nyd = nytimes_state_data()
drny = Arima_drop_state(ej, nyd)
drny
plot(drny)
opar = par(no.readonly=TRUE)
par(mar=c(4,3,2,2), mfrow=c(1,2))
plot(usa_full, main="all states", ylim=c(17000,38000))
plot(drny, main="excluding NY", ylim=c(17000,38000))
par(opar)
}
