% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arima.R
\name{Arima_nation}
\alias{Arima_nation}
\title{Use Rob Hyndman's forecast package to estimate drift in ARIMA models}
\usage{
Arima_nation(
  ejhu,
  alp3 = "USA",
  MAorder = 2,
  Difforder = 1,
  basedate = "2020-02-15",
  lookback_days = 29,
  ARorder = 0,
  max_date = NULL
)
}
\arguments{
\item{ejhu}{a tibble as returned by \code{enriched_jhu_data()}}

\item{alp3}{character(1) alpha3Code value for country}

\item{MAorder}{numeric(1) order of moving average component}

\item{Difforder}{numeric(1) differencing order d of ARIMA(p,d,q)}

\item{basedate}{character(1) used by lubridate::as_date to filter away all earlier records}

\item{lookback_days}{numeric(1) only uses this many days from most recent in ejhu}

\item{ARorder}{order of autoregressive component}

\item{max_date}{a date from which to start lookback ... defaults to NULL in which
case the latest available date is used}
}
\value{
instance of S3 class Arima_sars2pack
}
\description{
Use Rob Hyndman's forecast package to estimate drift in ARIMA models
}
\examples{
ej = enriched_jhu_data()
lkus = Arima_nation(ej)
lkus
plot(lkus)
}
