% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beoutbreakprepared.R
\name{beoutbreakprepared_data}
\alias{beoutbreakprepared_data}
\title{United States state and county level data from USAFacts}
\source{
https://github.com/beoutbreakprepared/nCoV2019
}
\usage{
beoutbreakprepared_data()
}
\value{

}
\description{
United States state and county level data from USAFacts
}
\note{

}
\examples{
res = beoutbreakprepared_data()
colnames(res)
head(res)

}
\seealso{
Other data-import: 
\code{\link{covidtracker_data}()},
\code{\link{jhu_data}()},
\code{\link{kff_icu_beds}()},
\code{\link{nytimes_county_data}()},
\code{\link{us_county_geo_details}()},
\code{\link{us_healthcare_capacity}()},
\code{\link{usa_facts_data}()}
}
\concept{data-import}
