% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_Rt.R
\name{bulk_estimate_Rt}
\alias{bulk_estimate_Rt}
\title{Estimate R(t) for all locales in a case-tracking dataset}
\usage{
bulk_estimate_Rt(
  df,
  grouping_columns = setdiff(colnames(df), c(cases_column, date_column)),
  cases_column = "count",
  date_column = "date",
  ...
)
}
\arguments{
\item{df}{a data.frame containing at least a date column and a cases column, describing
the cumulative cases at each date. The data.frame may contain additional columns
that can be used to "group" the dates and cases to produce a set of R(t) curves.}

\item{grouping_columns}{character() vector specifying the grouping columns that
will break \code{df} into chunks for estimation (i.e., location, etc.). The default is normally
correct and includes all
columns except for date_column and case_column.}

\item{cases_column}{character(1) the column in \code{df} that includes the case counts
of interest.}

\item{date_column}{character(1) the column in \code{df} that includes the date information
about when cases reported.}
}
\value{
A "stacked" data.frame with the outputs from the R(t) estimation process
and associated location information. The actual columns returned for the
R(t) estimate will depend on the \code{estimation_family} parameter as well
as other parameters specific to each method.
}
\description{
This function can estimate R(t) curves for all locations in a case-tracking
dataset and return a stacked data.frame with the location details included.
It is a convenience function for getting R(t) over a large dataset. Right
now, nothing is done in "parallel", so this function is not going to be
much (or any) faster than running on each location in a case-tracking
dataset independently and then combining.
}
\examples{
library(dplyr)
nyt = nytimes_state_data() \%>\% 
    dplyr::filter(subset=='confirmed') \%>\%
    dplyr::arrange(state,date)
head(nyt)

# this may produce warnings, but the processing
# will still be correct....
res = bulk_estimate_Rt(head(nyt,500), 
    estimation_family='epiestim',method = 'parametric_si',
    config = list(mean_si=3.96, std_si=4.75))
head(res)
colnames(res)

library(ggplot2)
ggplot(res, aes(x=date_start,y=1/`Mean(R)`,color=state)) + 
    geom_smooth(se=FALSE) + ylim(c(0,1.5))

}
\seealso{
Other analysis: 
\code{\link{estimate_Rt}()}

Other case-tracking: 
\code{\link{align_to_baseline}()},
\code{\link{beoutbreakprepared_data}()},
\code{\link{combined_us_cases_data}()},
\code{\link{covidtracker_data}()},
\code{\link{ecdc_data}()},
\code{\link{estimate_Rt}()},
\code{\link{eu_data_cache_data}()},
\code{\link{jhu_data}()},
\code{\link{nytimes_county_data}()},
\code{\link{plot_epicurve}()},
\code{\link{usa_facts_data}()}
}
\author{
Sean Davis \href{mailto:seandavi@gmail.com}{seandavi@gmail.com}
}
\concept{analysis}
\concept{case-tracking}
