% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cci_us_vaccine_data.R
\name{cci_us_vaccine_data}
\alias{cci_us_vaccine_data}
\title{Vaccine allocations and administrations in the United States}
\source{
\itemize{
\item \url{https://github.com/govex/COVID-19/tree/master/data_tables/vaccine_data}
}
}
\usage{
cci_us_vaccine_data()
}
\value{
A \code{data.frame}
}
\description{
This data resource captures vaccine allocations, doses administered,
and the number of people (as opposed to doses) to which they have
been administered. As of December, 2020, this dataset is not complete.
}
\examples{

vacc = cci_us_vaccine_data()
head(vacc)
table(vacc$unit)
table(vacc$status)
table(vacc$entity)

}
\seealso{
Other data-import: 
\code{\link{acaps_government_measures_data}()},
\code{\link{acaps_secondary_impact_data}()},
\code{\link{apple_mobility_data}()},
\code{\link{beoutbreakprepared_data}()},
\code{\link{cdc_aggregated_projections}()},
\code{\link{cdc_excess_deaths}()},
\code{\link{cdc_social_vulnerability_index}()},
\code{\link{coronadatascraper_data}()},
\code{\link{coronanet_government_response_data}()},
\code{\link{cov_glue_lineage_data}()},
\code{\link{cov_glue_newick_data}()},
\code{\link{cov_glue_snp_lineage}()},
\code{\link{covidtracker_data}()},
\code{\link{descartes_mobility_data}()},
\code{\link{ecdc_data}()},
\code{\link{economist_excess_deaths}()},
\code{\link{financial_times_excess_deaths}()},
\code{\link{google_mobility_data}()},
\code{\link{government_policy_timeline}()},
\code{\link{jhu_data}()},
\code{\link{jhu_us_data}()},
\code{\link{kff_icu_beds}()},
\code{\link{nytimes_county_data}()},
\code{\link{oecd_unemployment_data}()},
\code{\link{owid_data}()},
\code{\link{param_estimates_published}()},
\code{\link{test_and_trace_data}()},
\code{\link{us_county_geo_details}()},
\code{\link{us_county_health_rankings}()},
\code{\link{us_healthcare_capacity}()},
\code{\link{us_hospital_details}()},
\code{\link{us_state_distancing_policy}()},
\code{\link{usa_facts_data}()},
\code{\link{who_cases}()},
\code{\link{yougov_behavioral_data}()}
}
\author{
Sean Davis \href{mailto:seandavi@gmail.com}{seandavi@gmail.com}
}
\concept{data-import}
\concept{vaccines}
