% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdc_aggregated_projections.R
\name{cdc_aggregated_projections}
\alias{cdc_aggregated_projections}
\title{Projected deaths from COVID-19 models}
\usage{
cdc_aggregated_projections()
}
\description{
The US CDC gathers projections from several groups around
the world and aggregates them into a single data resource.
See the reference below for details of the models.
}
\details{
These models are not updated daily but more like weekly.
This function will attempt to grab the latest version.
}
\examples{
res = cdc_aggregated_projections()
head(res)
dplyr::glimpse(res)

# available models
table(res$model)

# projection targets
table(res$target)

min(res$forecast_date)
max(res$target_week_end_date)

}
\references{
\itemize{
\item \url{https://www.cdc.gov/coronavirus/2019-ncov/covid-data/forecasting-us.html}
}
}
\seealso{
Other data-import: 
\code{\link{apple_mobility_data}()},
\code{\link{beoutbreakprepared_data}()},
\code{\link{covidtracker_data}()},
\code{\link{descartes_mobility_data}()},
\code{\link{ecdc_data}()},
\code{\link{eu_data_cache_data}()},
\code{\link{google_mobility_data}()},
\code{\link{healthdata_projections_data}()},
\code{\link{jhu_data}()},
\code{\link{jhu_us_data}()},
\code{\link{kff_icu_beds}()},
\code{\link{nytimes_county_data}()},
\code{\link{us_county_geo_details}()},
\code{\link{us_healthcare_capacity}()},
\code{\link{us_state_distancing_policy}()},
\code{\link{usa_facts_data}()}
}
\author{
Sean Davis \href{mailto:seandavi@gmail.com}{seandavi@gmail.com}
}
\concept{data-import}
\concept{projections}
