% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdc_excess_deaths.R
\name{cdc_excess_deaths}
\alias{cdc_excess_deaths}
\title{US CDC excess deaths by week from 2018 to present}
\source{
\itemize{
\item \url{https://data.cdc.gov/NCHS/Excess-Deaths-Associated-with-COVID-19/xkkf-xrst}
}
}
\usage{
cdc_excess_deaths()
}
\description{
Understanding the potentially unmeasured death toll due to COVID-19
starts with understanding the expected vs observed death rates over
time. This dataset presents observed and expected deaths by state
by week of year for 2018 to present. In addition, the dataset can be
broken down into all cause death and that attributable to COVID-19
based on reported COVID-19 deaths.
}
\details{
Estimates of excess deaths can provide information about the burden
of mortality potentially related to COVID-19, beyond the number of
deaths that are directly attributed to COVID-19. Excess deaths are
typically defined as the difference between observed numbers of
deaths and expected numbers. This visualization provides weekly
data on excess deaths by jurisdiction of occurrence. Counts of
deaths in more recent weeks are compared with historical trends to
determine whether the number of deaths is significantly higher than
expected.

Estimates of excess deaths can be calculated in a variety of ways,
and will vary depending on the methodology and assumptions about
how many deaths are expected to occur. Estimates of excess deaths
presented in this webpage were calculated using Farrington
surveillance algorithms (1). For each jurisdiction, a model is used
to generate a set of expected counts, and the upper bound of the
95\% Confidence Intervals (95\% CI) of these expected counts is used
as a threshold to estimate excess deaths. Observed counts are
compared to these upper bound estimates to determine whether a
significant increase in deaths has occurred. Provisional counts are
weighted to account for potential underreporting in the most recent
weeks. However, data for the most recent week(s) are still likely
to be incomplete. Only about 60\% of deaths are reported within 10
days of the date of death, and there is considerable variation by
jurisdiction. More detail about the methods, weighting, data, and
limitations can be found in the \href{https://www.cdc.gov/nchs/nvss/vsrr/covid19/excess_deaths.htm}{Technical Notes}.
}
\examples{
cdcdeaths = cdc_excess_deaths()
head(cdcdeaths)
colnames(cdcdeaths)
table(cdcdeaths$outcome)


# Examine excess deaths in three states
library(ggplot2)
library(dplyr)
interesting_states = c('Ohio', 'Pennsylvania', 'California')
ggplot(cdcdeaths \%>\% dplyr::filter(type=="Predicted (weighted)" &
                            outcome=="All causes" &
                            state \%in\% interesting_states),
        aes(x=date,y=deaths)) +
    geom_point(aes(color=deaths>upper_bound_threshold)) +
    geom_line(aes(x=date,y=upper_bound_threshold)) +
    facet_grid(rows=vars(state)) +
    ggtitle('Excess deaths over time')

}
\references{
\itemize{
\item \url{https://www.cdc.gov/nchs/nvss/vsrr/covid19/excess_deaths.htm}
}
}
\seealso{
Other data-import: 
\code{\link{apple_mobility_data}()},
\code{\link{beoutbreakprepared_data}()},
\code{\link{cdc_aggregated_projections}()},
\code{\link{coronanet_government_response_data}()},
\code{\link{cov_glue_lineage_data}()},
\code{\link{cov_glue_newick_data}()},
\code{\link{covidtracker_data}()},
\code{\link{descartes_mobility_data}()},
\code{\link{ecdc_data}()},
\code{\link{eu_data_cache_data}()},
\code{\link{google_mobility_data}()},
\code{\link{government_policy_timeline}()},
\code{\link{healthdata_projections_data}()},
\code{\link{jhu_data}()},
\code{\link{jhu_us_data}()},
\code{\link{kff_icu_beds}()},
\code{\link{nytimes_county_data}()},
\code{\link{param_estimates_published}()},
\code{\link{us_county_geo_details}()},
\code{\link{us_county_health_rankings}()},
\code{\link{us_healthcare_capacity}()},
\code{\link{us_hospital_details}()},
\code{\link{us_state_distancing_policy}()},
\code{\link{usa_facts_data}()}
}
\author{
Sean Davis \href{mailto:seandavi@gmail.com}{seandavi@gmail.com}
}
\concept{data-import}
