% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_summaries.R
\name{column_summaries}
\alias{column_summaries}
\title{convenience function for summarizing data.frame columns}
\usage{
column_summaries(x, dates_as_char = TRUE)
}
\arguments{
\item{dates_as_char}{logical() whether to treat date columns
as character}

\item{df}{a data.frame object}
}
\value{
a \code{tibble} with two columns, \code{name} and \code{column_details}.
The name column is the column name. The \code{column_details}
column is a list with five elements:
\itemize{
\item min: the minumum value (including for dates and character)
\item max: the maximum value (including for dates and character)
\item class: the class of the column
\item sample_values: up to five randomly-sampled values
}
}
\description{
This function takes a data.frame and returns a simple
summary of the columns and their content as a data.frame.
}
\examples{
data(iris)
cs = column_summaries(iris)
head(cs)
str(cs$column_details)

}
