% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_us_cases_data.R
\name{combined_us_cases_data}
\alias{combined_us_cases_data}
\title{gather US states confirmed cases from multiple sources}
\usage{
combined_us_cases_data()
}
\description{
This is a convenience function to combine results from
sources providing US state-level datasets. A primary use
case is to provide fodder for comparison and quality control
over all the available datasets.
}
\examples{
cus = combined_us_cases_data()
head(cus)
unique(cus$dataset)
plot_epicurve(cus, filter_expression = state == 'GA',color='dataset') +
    ggtitle('Cumulative confirmed cases in Georgia')

plot_epicurve(cus, filter_expression = state == 'GA' & incidence>10,
    case_column='incidence', color='dataset', log=FALSE) +
    ggtitle('Daily confirmed cases in Georgia') + geom_smooth(alpha=0.25)

}
\seealso{
Other case-tracking: 
\code{\link{align_to_baseline}()},
\code{\link{beoutbreakprepared_data}()},
\code{\link{bulk_estimate_Rt}()},
\code{\link{covidtracker_data}()},
\code{\link{ecdc_data}()},
\code{\link{estimate_Rt}()},
\code{\link{eu_data_cache_data}()},
\code{\link{jhu_data}()},
\code{\link{nytimes_county_data}()},
\code{\link{plot_epicurve}()},
\code{\link{usa_facts_data}()}
}
\author{
Sean Davis \href{mailto:seandavi@gmail.com}{seandavi@gmail.com}
}
\concept{case-tracking}
