% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arima.R
\name{contig_states_twolet}
\alias{contig_states_twolet}
\title{provide vector of contiguous US state names and DC, abbr}
\usage{
contig_states_twolet()
}
\description{
provide vector of contiguous US state names and DC, abbr
}
