% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coronadatascraper_data.R
\name{coronadatascraper_data}
\alias{coronadatascraper_data}
\title{Time series cases, hospitalizations, and testing from Corona Data Scraper project}
\source{
\url{https://coronadatascraper.com/timeseries.csv}
}
\usage{
coronadatascraper_data()
}
\description{
Corona Data Scraper pulls COVID-19 Coronavirus case data from
verified sources, finds the corresponding GeoJSON features, and
adds population data. All sources are cited right in the same row
as the data. Covers testing, hospitalization, deaths, cases, icu.
Includes population and geospatial coordinates.
}
\examples{
res = coronadatascraper_data()
colnames(res)
head(res)
dplyr::glimpse(res)

length(unique(res$names))
length(unique(res$county))
length(unique(res$state))
length(unique(res$country))

}
\references{
\url{https://coronadatascraper.com/#home}
}
\seealso{
Other data-import: 
\code{\link{acaps_government_measures_data}()},
\code{\link{apple_mobility_data}()},
\code{\link{beoutbreakprepared_data}()},
\code{\link{cdc_aggregated_projections}()},
\code{\link{cdc_excess_deaths}()},
\code{\link{cdc_social_vulnerability_index}()},
\code{\link{coronanet_government_response_data}()},
\code{\link{cov_glue_lineage_data}()},
\code{\link{cov_glue_newick_data}()},
\code{\link{cov_glue_snp_lineage}()},
\code{\link{covidtracker_data}()},
\code{\link{descartes_mobility_data}()},
\code{\link{ecdc_data}()},
\code{\link{economist_excess_deaths}()},
\code{\link{financial_times_excess_deaths}()},
\code{\link{google_mobility_data}()},
\code{\link{government_policy_timeline}()},
\code{\link{jhu_data}()},
\code{\link{jhu_us_data}()},
\code{\link{kff_icu_beds}()},
\code{\link{nytimes_county_data}()},
\code{\link{oecd_unemployment_data}()},
\code{\link{owid_data}()},
\code{\link{param_estimates_published}()},
\code{\link{test_and_trace_data}()},
\code{\link{us_county_geo_details}()},
\code{\link{us_county_health_rankings}()},
\code{\link{us_healthcare_capacity}()},
\code{\link{us_hospital_details}()},
\code{\link{us_state_distancing_policy}()},
\code{\link{usa_facts_data}()},
\code{\link{who_cases}()},
\code{\link{yougov_behavioral_data}()}

Other case-tracking: 
\code{\link{align_to_baseline}()},
\code{\link{beoutbreakprepared_data}()},
\code{\link{bulk_estimate_Rt}()},
\code{\link{combined_us_cases_data}()},
\code{\link{covidtracker_data}()},
\code{\link{ecdc_data}()},
\code{\link{estimate_Rt}()},
\code{\link{jhu_data}()},
\code{\link{nytimes_county_data}()},
\code{\link{owid_data}()},
\code{\link{plot_epicurve}()},
\code{\link{test_and_trace_data}()},
\code{\link{usa_facts_data}()},
\code{\link{who_cases}()}
}
\author{
Sean Davis \href{mailto:seandavi@gmail.com}{seandavi@gmail.com}
}
\concept{case-tracking}
\concept{data-import}
