% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coronanet_government_response_data.R
\name{coronanet_government_response_data}
\alias{coronanet_government_response_data}
\title{CoronaNet government policy response database}
\source{
\itemize{
\item \url{https://coronanet-project.org/download}
}
}
\usage{
coronanet_government_response_data()
}
\value{
\itemize{
\item record_id Unique identifier for each policy record
\item entry_type Whether the record is new, meaning no restriction had been in place before, or an update (restriction was in place but changed). Corrections are corrections to previous entries.
\item event_description A short description of the policy change
\item type The category of the policy
\item country The country initiating the policy
\item init_country_level Whether the policy came from the national level or a sub-national unit
\item index_prov The ID of the sub-national unit
\item target_country Which foreign country a policy is targeted at (i.e. travel policies)
\item target_geog_level Whether the target of the policy is a country as a whole or a sub-national unit of that country
\item target_who_what Who the policy is targeted at
\item recorded_date When the record was entered into our data
\item target_direction Whether a travel-related policy affects people coming in (Inbound) or leaving (Outbound)
\item travel_mechanism If a travel policy, what kind of transportation it affects
\item compliance Whether the policy is voluntary or mandatory
\item enforcer What unit in the country is responsible for enforcement
\item date_announced When the policy goes into effect
\item link A link to at least one source for the policy
\item ISO_A3 3-digit ISO country codes
\item ISO_A2 2-digit ISO country codes
\item severity_index_5perc 5\% posterior low estimate (i.e. lower bound of uncertainty interval) for severity index
\item severity_index_median posterior median estimate (point estimate) for severity index, which comes from a Bayesian latent variable model aggregating across policy types to measure country-level policy severity (see paper on our website)
\item severity_index_5perc 95\% posterior high estimate (i.e. upper bound of uncertainty interval) for severity index
}
}
\description{
This dataset contains variables from the CoronaNet government
response project, representing national and sub-national policy
event data from more than 140 countries since January 1st,
2020. The data include source links, descriptions, targets
(i.e. other countries), the type and level of enforcement, and a
comprehensive set of policy types.
}
\examples{
res = coronanet_government_response_data()
head(res)
colnames(res)
dplyr::glimpse(res)
summary(res)


}
\references{
Cheng, Cindy, Joan Barcelo, Allison Hartnett,
Robert Kubinec, and Luca Messerschmidt. 2020. “Coronanet: A
Dyadic Dataset of Government Responses to the COVID-19
Pandemic.” SocArXiv. April 12. doi:10.31235/osf.io/dkvxy.
\itemize{
\item \url{https://coronanet-project.org/working_papers}
}
}
\seealso{
Other data-import: 
\code{\link{apple_mobility_data}()},
\code{\link{beoutbreakprepared_data}()},
\code{\link{cdc_aggregated_projections}()},
\code{\link{cdc_excess_deaths}()},
\code{\link{cov_glue_lineage_data}()},
\code{\link{cov_glue_newick_data}()},
\code{\link{covidtracker_data}()},
\code{\link{descartes_mobility_data}()},
\code{\link{ecdc_data}()},
\code{\link{eu_data_cache_data}()},
\code{\link{google_mobility_data}()},
\code{\link{government_policy_timeline}()},
\code{\link{healthdata_projections_data}()},
\code{\link{jhu_data}()},
\code{\link{jhu_us_data}()},
\code{\link{kff_icu_beds}()},
\code{\link{nytimes_county_data}()},
\code{\link{param_estimates_published}()},
\code{\link{us_county_geo_details}()},
\code{\link{us_county_health_rankings}()},
\code{\link{us_healthcare_capacity}()},
\code{\link{us_hospital_details}()},
\code{\link{us_state_distancing_policy}()},
\code{\link{usa_facts_data}()}

Other NPI: 
\code{\link{government_policy_timeline}()}
}
\concept{NPI}
\concept{data-import}
