% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_metadata.R
\name{country_metadata}
\alias{country_metadata}
\title{Get basic country data}
\usage{
country_metadata()
}
\value{
a data.frame
}
\description{
This function uses the rest api here
\url{https://restcountries.eu/rest/v2/all'}
to get country metadata.
}
\examples{
cmd = country_metadata()
colnames(cmd)
sapply(cmd, class)

head(cmd)

}
\seealso{
\url{https://restcountries.eu/rest/v2/all}
}
\author{
Sean Davis \href{mailto:seandavi@gmail.com}{seandavi@gmail.com}
}
