% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_glue_lineage_data.R
\name{cov_glue_snp_lineage}
\alias{cov_glue_snp_lineage}
\title{COVID GLUE snp to lineage and taxon map}
\usage{
cov_glue_snp_lineage()
}
\description{
This function maps lineage name to snps and to
taxon names. Everything is GISAID-specific.
}
\examples{
res = cov_glue_snp_lineage()
res
colnames(res)
dplyr::glimpse(res)

}
\references{
\itemize{
\item \url{https://www.biorxiv.org/content/10.1101/2020.04.17.046086v1}
}
}
\seealso{
cov_glue_lineage_data(), cov_glue_newick_data()

Other lineage: 
\code{\link{cov_glue_lineage_data}()},
\code{\link{cov_glue_newick_data}()}

Other data-import: 
\code{\link{acaps_government_measures_data}()},
\code{\link{apple_mobility_data}()},
\code{\link{beoutbreakprepared_data}()},
\code{\link{cdc_aggregated_projections}()},
\code{\link{cdc_excess_deaths}()},
\code{\link{cdc_social_vulnerability_index}()},
\code{\link{coronadatascraper_data}()},
\code{\link{coronanet_government_response_data}()},
\code{\link{cov_glue_lineage_data}()},
\code{\link{cov_glue_newick_data}()},
\code{\link{covidtracker_data}()},
\code{\link{descartes_mobility_data}()},
\code{\link{ecdc_data}()},
\code{\link{economist_excess_deaths}()},
\code{\link{eu_data_cache_data}()},
\code{\link{financial_times_excess_deaths}()},
\code{\link{google_mobility_data}()},
\code{\link{government_policy_timeline}()},
\code{\link{jhu_data}()},
\code{\link{jhu_us_data}()},
\code{\link{kff_icu_beds}()},
\code{\link{nytimes_county_data}()},
\code{\link{oecd_unemployment_data}()},
\code{\link{owid_data}()},
\code{\link{param_estimates_published}()},
\code{\link{test_and_trace_data}()},
\code{\link{us_county_geo_details}()},
\code{\link{us_county_health_rankings}()},
\code{\link{us_healthcare_capacity}()},
\code{\link{us_hospital_details}()},
\code{\link{us_state_distancing_policy}()},
\code{\link{usa_facts_data}()},
\code{\link{who_cases}()},
\code{\link{yougov_behavioral_data}()}
}
\concept{data-import}
\concept{lineage}
