% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_covact.R
\name{covact_states_data}
\alias{covact_states_data}
\title{use methods described at \verb{https://apidocs.covidactnow.org/} to acquire time series at state level}
\usage{
covact_states_data(key = Sys.getenv("COVACT_KEY"), quiet = TRUE)
}
\arguments{
\item{key}{character(1) 32-character key.  Default value is
retrieved from environment variable \code{COVACT_KEY}.  See Note.}

\item{quiet}{logical(1) defaults to TRUE, in which case the messaging of tibble ingestion is
suppressed.}
}
\value{
a tibble with one line per day per state
}
\description{
use methods described at \verb{https://apidocs.covidactnow.org/} to acquire time series at state level
}
\note{
Acquire an API key at \verb{https://apidocs.covidactnow.org/}.  There
is a nontrivial delay between the delivery of key (immediate) and
recognition of the key by the service.  You can place the
line \code{COVACT_KEY=...} in your .Renviron file to make this
value available for every normal R session.
}
