% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covact_vbls}
\alias{covact_vbls}
\title{covactnow state-level snapshot, data variable summary}
\format{
data.frame
}
\usage{
covact_vbls
}
\description{
covactnow state-level snapshot, data variable summary
}
\keyword{datasets}
