% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidtracking_com.R
\name{covidtracker_data}
\alias{covidtracker_data}
\title{US State-level data including test results}
\usage{
covidtracker_data()
}
\value{
A tidy \code{tbl_df}
}
\description{
Get data from the COVID tracking project
\url{https://covidtracking.com/} including daily historical data on
testing results, hospitalizations, and deaths.
}
\details{
From the COVID Tracking Project website:

The COVID Tracking Project collects information from 50 US states,
the District of Columbia, and 5 other US territories to provide the
most comprehensive testing data we can collect for the novel
coronavirus, SARS-CoV-2. We attempt to report positive and negative
results, pending tests, and total people tested for each state or
district currently reporting that data.
}
\examples{
res = covidtracker_data()
colnames(res)
dim(res)
dplyr::glimpse(res)


}
\seealso{
\itemize{
\item \url{https://covidtracking.com/about-tracker/}
\item \url{https://covidtracking.com/notes/}
}

Other data-import: 
\code{\link{apple_mobility_data}()},
\code{\link{beoutbreakprepared_data}()},
\code{\link{cdc_aggregated_projections}()},
\code{\link{descartes_mobility_data}()},
\code{\link{ecdc_data}()},
\code{\link{eu_data_cache_data}()},
\code{\link{google_mobility_data}()},
\code{\link{healthdata_projections_data}()},
\code{\link{jhu_data}()},
\code{\link{jhu_us_data}()},
\code{\link{kff_icu_beds}()},
\code{\link{nytimes_county_data}()},
\code{\link{us_county_geo_details}()},
\code{\link{us_healthcare_capacity}()},
\code{\link{us_state_distancing_policy}()},
\code{\link{usa_facts_data}()}
}
\concept{data-import}
