% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_catalog.R
\name{create_dataset_details}
\alias{create_dataset_details}
\title{create data details yaml file for all datasets}
\usage{
create_dataset_details(fname = "inst/data_catalog/dataset_details.yaml")
}
\description{
This function runs available_datasets() and then
loops over the datasets to collect a list that looks like:
}
\details{
\preformatted{kff_icu_beds:
columns:
    fips: character
county: character
st: character
state: character
hospitals_in_cost_reports: numeric
Hospitals_in_HC: numeric
all_icu: numeric
Total_pop: numeric
60plus: numeric
60plus_pct: numeric
60plus_per_each_icu_bed: numeric
dimensions:
    nrow: 3142
ncol: 11
us_county_geo_details:
    columns:
    state: character
fips: character
ansicode: character
county: character
area_land: numeric
area_water: numeric
area_land_sqmi: numeric
area_water_sqmi: numeric
geometry:
    - sfc_POINT
    - sfc
dimensions:
    nrow: 3220
    ncol: 9
}
}
\keyword{internal}
