% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CRF_468_China}
\alias{CRF_468_China}
\title{serial interval values fo 468 COVID-19 cases in 93 cities in China}
\format{
data.frame
}
\source{
\url{https://github.com/MeyersLabUTexas/COVID-19} Table S5.xlsx
}
\usage{
CRF_468_China
}
\description{
serial interval values fo 468 COVID-19 cases in 93 cities in China
}
\keyword{datasets}
