% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejhu.R
\name{cumulative_events_ejhu}
\alias{cumulative_events_ejhu}
\title{constructor for covid_events}
\usage{
cumulative_events_ejhu(
  src,
  eventtype = "confirmed",
  alpha3 = "USA",
  ProvinceState = NULL
)
}
\arguments{
\item{src}{as retrieved with enhanced_jhu_data}

\item{eventtype}{character(1) 'confirmed' or 'deaths'}

\item{alpha3}{character(1) code for country}

\item{ProvinceState}{character(1) for province, default to NULL}
}
\description{
constructor for covid_events
}
