% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descartes_mobility.R
\name{decartes_mobility_data}
\alias{decartes_mobility_data}
\title{Get US daily mobility data from Descarte Labs}
\source{
\itemize{
\item https://raw.githubusercontent.com/descarteslabs/DL-COVID-19/master/DL-us-mobility-daterow.csv
}
}
\usage{
decartes_mobility_data()
}
\value{
\itemize{
\item country_code: ISO 3166-1 alpha-2 code.
\item admin_level: 0 for country, 1 for admin1, 2 for admin2 granularity.
\item admin1: GeoNames ADM1 feature name for the first-order administrative division, such as a state in the United States.
\item admin2: GeoNames ADM2 feature name for the second-order administrative division, such as a county or borough in the United States.
\item fips: FIPS code, a standard geographic identifier, to make it easier to combine this data with other data sets.
\item samples: The number of samples observed in the specified region.
\item m50: The median of the max-distance mobility for all samples in the specified region.
\item m50_index: The percent of normal m50 in the region, with normal m50 defined during 2020-02-17 to 2020-03-07.
}
}
\description{
Get US daily mobility data from Descarte Labs
}
\examples{
res = descartes_mobility_data()
colnames(res)
glimpse(res)

}
\seealso{
Other data-import: 
\code{\link{beoutbreakprepared_data}()},
\code{\link{covidtracker_data}()},
\code{\link{healthdata_projections_data}()},
\code{\link{jhu_data}()},
\code{\link{jhu_us_data}()},
\code{\link{kff_icu_beds}()},
\code{\link{nytimes_county_data}()},
\code{\link{us_county_geo_details}()},
\code{\link{us_healthcare_capacity}()},
\code{\link{usa_facts_data}()}
}
\author{
Sean Davis \href{mailto:seandavi@gmail.com}{seandavi@gmail.com}
}
\concept{data-import}
