% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdemo.R
\name{demo_simulation}
\alias{demo_simulation}
\title{set up a simulation as demonstrated by John C Mallery}
\usage{
demo_simulation(
  gt.type = "gamma",
  gt.val = c(3, 1.5),
  begin = 1,
  end = 30,
  epid.nb = 1,
  epid.length = 30,
  family = "poisson",
  R0 = 1.67,
  peak.value = 500
)
}
\arguments{
\item{gt.type}{character(1) \code{type} setting for \code{R0:generation.time}}

\item{gt.val}{numeric() \code{val} setting for \code{R0:generation.time}}

\item{begin}{numeric(1) passed to \code{R0::estimate.R}}

\item{end}{numeric(1) passed to \code{R0::estimate.R}}

\item{epid.nb}{numeric(1) passed to \code{R0:sim.epid}}

\item{epid.length}{numeric(1) passed to \code{R0:sim.epid}}

\item{family}{character(1) passed to \code{R0:sim.epid}}

\item{R0}{numeric(1) passed to \code{R0:sim.epid}}

\item{peak.value}{numeric(1) passed to \code{R0:sim.epid}}
}
\value{
result of \code{estimate.R}
}
\description{
set up a simulation as demonstrated by John C Mallery
}
\examples{
\dontrun{
d1 = demo_simulation()
# plot(d1) ## note: R0:::plot.R0.sR is very peculiar and runs dev.new()
plotfit(d1)  # same here but seems to have simpler behavior
# s.a <- sensitivity.analysis(res=d1$estimates$EG, begin=1:15, end=16:30, sa.type="time")
}
}
