% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descartes_mobility.R
\name{descartes_mobility_data}
\alias{descartes_mobility_data}
\title{Get US daily mobility data from Descartes Labs}
\source{
\itemize{
\item \url{https://raw.githubusercontent.com/descarteslabs/DL-COVID-19/master/DL-us-mobility-daterow.csv}
}
}
\usage{
descartes_mobility_data()
}
\value{
\itemize{
\item country_code: ISO 3166-1 alpha-2 code.
\item admin_level: 0 for country, 1 for admin1, 2 for admin2 granularity.
\item admin1: GeoNames ADM1 feature name for the first-order administrative division, such as a state in the United States.
\item admin2: GeoNames ADM2 feature name for the second-order administrative division, such as a county or borough in the United States.
\item fips: FIPS code, a standard geographic identifier, to make it easier to combine this data with other data sets.
\item samples: The number of samples observed in the specified region.
\item m50: The median of the max-distance mobility for all samples in the specified region.
\item m50_index: The percent of normal m50 in the region, with normal m50 defined during 2020-02-17 to 2020-03-07.
}
}
\description{
Descartes Labs quantify the level of human mobility in the US at
the state level. Their methodology looks at a collection of mobile
devices reporting consistently throughout the day. They calculate the
maximum distance moved in kilometers (excluding outliers) from the
first reported location. Using this value, they calculate the median
across all devices in the sample to generate a mobility metric for
selected states.
}
\section{License}{


Creative Commons Attribution (CC BY 4.0), see \url{https://github.com/descarteslabs/DL-COVID-19}
}

\examples{
library(dplyr)
res = descartes_mobility_data()
colnames(res)
dplyr::glimpse(res)

# plot data for Georgia
library(ggplot2)

# this gets us state-level data
GA = res \%>\% dplyr::filter(admin1=='Georgia' & admin_level==1)

ggplot(GA, aes(x=date, y=m50_index)) + geom_line() + ggtitle('M50 index over time in Georgia')

#limit to dates of interest around time that GA is reopening

GA = GA \%>\% dplyr::filter(date>as.Date('2020-01-01') & date < as.Date('2020-06-01'))
ggplot(GA, aes(x=date, y=m50_index)) +
    geom_line() +
    ggtitle('M50 index over time in Georgia')
# Obviously, there are day-specific effects, so use
# R to "regress out" the effects of day of week to better
# observe trend.
GA = GA \%>\% dplyr::mutate(dow = lubridate::wday(date, label=TRUE))
lmfit = lm(m50_index ~ dow, data = GA)

# We are interested in the residuals, or
# the variation in the data not explained by
# the day of the week.
GA$m50_index_regressed = residuals(lmfit)

ggplot(GA, aes(x=date, y=m50_index_regressed)) +
    geom_smooth() +
    geom_point()

# Compare states
states = c("New York", "California", "Nevada",
           "Texas", "Georgia", "Florida")
ST = res \%>\%
          dplyr::filter(admin1 \%in\% states & admin_level==1) \%>\%
          dplyr::mutate(dow = lubridate::wday(date, label=TRUE))
lmfit = lm(m50_index ~ dow + admin1, data = ST)
ST$m50_index_regressed = residuals(lmfit)

ggplot(ST, aes(x=date, y=m50_index_regressed)) +
    geom_smooth() +
    geom_point() +
    facet_wrap('admin1', nrow=2) +
    ggtitle('Mobility index differences across states')



}
\references{
\itemize{
\item \url{https://www.descarteslabs.com/mobility-v097/}
\item \url{https://www.descarteslabs.com/wp-content/uploads/2020/03/mobility-v097.pdf}
}
}
\seealso{
Other data-import: 
\code{\link{apple_mobility_data}()},
\code{\link{beoutbreakprepared_data}()},
\code{\link{cdc_aggregated_projections}()},
\code{\link{covidtracker_data}()},
\code{\link{ecdc_data}()},
\code{\link{eu_data_cache_data}()},
\code{\link{google_mobility_data}()},
\code{\link{government_policy_timeline}()},
\code{\link{healthdata_projections_data}()},
\code{\link{jhu_data}()},
\code{\link{jhu_us_data}()},
\code{\link{kff_icu_beds}()},
\code{\link{nytimes_county_data}()},
\code{\link{param_estimates_published}()},
\code{\link{us_county_geo_details}()},
\code{\link{us_healthcare_capacity}()},
\code{\link{us_state_distancing_policy}()},
\code{\link{usa_facts_data}()}

Other mobility: 
\code{\link{apple_mobility_data}()},
\code{\link{google_mobility_data}()}
}
\author{
Sean Davis \href{mailto:seandavi@gmail.com}{seandavi@gmail.com}
}
\concept{data-import}
\concept{mobility}
