% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econ_tracker_employment.R
\name{.melt_econ_tracker_employment}
\alias{.melt_econ_tracker_employment}
\title{Convert employment data to long form}
\usage{
.melt_econ_tracker_employment(dtable)
}
\arguments{
\item{dtable}{a \code{data.table} object}
}
\value{
a long-form \code{data.table}
}
\description{
Also adds columns to make data easier to
filter.
}
\keyword{internal}
