% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jhu_data.R
\name{.munge_data_from_jhu}
\alias{.munge_data_from_jhu}
\title{simple function to munge JHU data into long-form tibble}
\usage{
.munge_data_from_jhu(subset)
}
\arguments{
\item{subset}{character(1) of Confirmed, Deaths, Recovered}
}
\value{
a long-form tibble
}
\description{
This function takes one of three subsets--confirmed,
deaths, recovered--and munges.
}
\keyword{internal}
