% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyEG.R, R/easyRt.R
\name{easyEG_eucache}
\alias{easyEG_eucache}
\title{simplify production of EG fit using eu_data_cache_data}
\usage{
easyEG_eucache(gt, filt, src, start_date, end_date)

easyEG_eucache(gt, filt, src, start_date, end_date)
}
\arguments{
\item{gt}{output of R0::generation.time}

\item{filt}{an expression selecting a region, like nuts_3 == "Bergamo"}

\item{src}{instance of eu_data_cache_data(), which will be run if this is missing}

\item{start_date}{string in form yyyy-mm-dd}

\item{end_date}{string in form yyyy-mm-dd}
}
\description{
simplify production of EG fit using eu_data_cache_data

simplify production of Rt fit using eu_data_cache_data
}
\examples{
gt = R0::generation.time("gamma", c(3.5, 4.8))
run1 = easyEG_eucache(gt, nuts_3=="Bergamo", start_date="2020-02-20", 
  end_date="2020-03-15")
plot(run1)
est_R0(run1)
low_R0(run1)
hi_R0(run1)
gt = R0::generation.time("gamma", c(3.5, 4.8))
run1 = easyEG_eucache(gt, nuts_3=="Bergamo", start_date="2020-02-20", 
  end_date="2020-03-15")
plot(run1)
est_R0(run1)
low_R0(run1)
hi_R0(run1)
}
