% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdc_data.R
\name{ecdc_data}
\alias{ecdc_data}
\title{Daily COVID-19 deaths and confirmed cases from the European CDC}
\source{
\itemize{
\item \url{https://www.ecdc.europa.eu/en/covid-19}
}
}
\usage{
ecdc_data()
}
\description{
Daily COVID-19 deaths and confirmed cases from the European CDC
}
\details{
From the \href{https://www.ecdc.europa.eu/en/covid-19/data-collection}{ECDC website}
last accessed April 24, 2020:

Since the beginning of the coronavirus pandemic, ECDC’s Epidemic
Intelligence team has been collecting the number of COVID-19 cases
and deaths, based on reports from health authorities
worldwide. This comprehensive and systematic process is carried out
on a daily basis. To insure the accuracy and reliability of the
data, this process is being constantly refined. This helps to
monitor and interpret the dynamics of the COVID-19 pandemic not
only in the European Union (EU), the European Economic Area (EEA),
but also worldwide.

Every day between 6.00 and 10.00 CET, a team of epidemiologists
screens up to 500 relevant sources to collect the latest
figures. The data screening is followed by ECDC’s standard epidemic
intelligence process for which every single data entry is validated
and documented in an ECDC database. An extract of this database,
complete with up-to-date figures and data visualisations, is then
shared on the ECDC website, ensuring a maximum level of
transparency.

ECDC receives regular updates from EU/EEA countries through the
Early Warning and Response System (EWRS), The European Surveillance
System (TESSy), the World Health Organization (WHO) and email
exchanges with other international stakeholders. This information
is complemented by screening up to 500 sources every day to collect
COVID-19 figures from 196 countries. This includes websites of
ministries of health (43\% of the total number of sources), websites
of public health institutes (9\%), websites from other national
authorities (ministries of social services and welfare,
governments, prime minister cabinets, cabinets of ministries,
websites on health statistics and official response teams) (6\%),
WHO websites and WHO situation reports (2\%), and official
dashboards and interactive maps from national and international
institutions (10\%). In addition, ECDC screens social media accounts
maintained by national authorities, for example Twitter, Facebook,
YouTube or Telegram accounts run by ministries of health (28\%) and
other official sources (e.g. official media outlets) (2\%). Several
media and social media sources are screened to gather additional
information which can be validated with the official sources
previously mentioned. Only cases and deaths reported by the
national and regional competent authorities from the countries and
territories listed are aggregated in our database.
}
\examples{
ecdc = ecdc_data()
colnames(ecdc)
dplyr::glimpse(ecdc)

}
\seealso{
Other data-import: 
\code{\link{apple_mobility_data}()},
\code{\link{beoutbreakprepared_data}()},
\code{\link{cdc_aggregated_projections}()},
\code{\link{cdc_excess_deaths}()},
\code{\link{coronanet_government_response_data}()},
\code{\link{cov_glue_lineage_data}()},
\code{\link{cov_glue_newick_data}()},
\code{\link{covidtracker_data}()},
\code{\link{descartes_mobility_data}()},
\code{\link{eu_data_cache_data}()},
\code{\link{google_mobility_data}()},
\code{\link{government_policy_timeline}()},
\code{\link{healthdata_projections_data}()},
\code{\link{jhu_data}()},
\code{\link{jhu_us_data}()},
\code{\link{kff_icu_beds}()},
\code{\link{nytimes_county_data}()},
\code{\link{param_estimates_published}()},
\code{\link{us_county_geo_details}()},
\code{\link{us_county_health_rankings}()},
\code{\link{us_healthcare_capacity}()},
\code{\link{us_hospital_details}()},
\code{\link{us_state_distancing_policy}()},
\code{\link{usa_facts_data}()}

Other data-import: 
\code{\link{apple_mobility_data}()},
\code{\link{beoutbreakprepared_data}()},
\code{\link{cdc_aggregated_projections}()},
\code{\link{cdc_excess_deaths}()},
\code{\link{coronanet_government_response_data}()},
\code{\link{cov_glue_lineage_data}()},
\code{\link{cov_glue_newick_data}()},
\code{\link{covidtracker_data}()},
\code{\link{descartes_mobility_data}()},
\code{\link{eu_data_cache_data}()},
\code{\link{google_mobility_data}()},
\code{\link{government_policy_timeline}()},
\code{\link{healthdata_projections_data}()},
\code{\link{jhu_data}()},
\code{\link{jhu_us_data}()},
\code{\link{kff_icu_beds}()},
\code{\link{nytimes_county_data}()},
\code{\link{param_estimates_published}()},
\code{\link{us_county_geo_details}()},
\code{\link{us_county_health_rankings}()},
\code{\link{us_healthcare_capacity}()},
\code{\link{us_hospital_details}()},
\code{\link{us_state_distancing_policy}()},
\code{\link{usa_facts_data}()}

Other case-tracking: 
\code{\link{align_to_baseline}()},
\code{\link{beoutbreakprepared_data}()},
\code{\link{bulk_estimate_Rt}()},
\code{\link{combined_us_cases_data}()},
\code{\link{covidtracker_data}()},
\code{\link{estimate_Rt}()},
\code{\link{eu_data_cache_data}()},
\code{\link{jhu_data}()},
\code{\link{nytimes_county_data}()},
\code{\link{owid_data}()},
\code{\link{plot_epicurve}()},
\code{\link{usa_facts_data}()}
}
\author{
Sean Davis \href{mailto:seandavi@gmail.com}{seandavi@gmail.com}
}
\concept{case-tracking}
\concept{data-import}
