% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econ_tracker_consumer_spending.R
\name{econ_tracker_consumer_spending}
\alias{econ_tracker_consumer_spending}
\alias{econ_tracker_consumer_spending_city_data}
\alias{econ_tracker_consumer_spending_county_data}
\alias{econ_tracker_consumer_spending_state_data}
\alias{econ_tracker_consumer_spending_national_data}
\title{United States Consumer Spending}
\source{
\itemize{
\item Affinity Solutions via Opportunity Insight econ tracker
\item Update Frequency: Weekly
}
}
\usage{
econ_tracker_consumer_spending_city_data()

econ_tracker_consumer_spending_county_data()

econ_tracker_consumer_spending_state_data()

econ_tracker_consumer_spending_national_data()
}
\description{
Aggregated and anonymized purchase data from consumer credit and debit card spending. Spending is reported based on the ZIP code where the cardholder lives, not the ZIP code where transactions occurred.
}
\details{
Data are Seasonally adjusted change since January 2020. Data is indexed in 2019 and 2020 as the change relative to the January index period. We then seasonally adjust by dividing year-over-year, which represents the difference between the change since January observed in 2020 compared to the change since January observed since 2019. We account for differences in the dates of federal holidays between 2019 and 2020 by shifting the 2019 reference data to align the holidays before performing the year-over-year division.

Geographies: National, State, County, Metro
\subsection{Breakdowns:}{
\subsection{By Industry. Industries are constructed by grouping merchant codes that are used by Affinity Solutions to identify the category of merchant and merchant activity.}{
\itemize{
\item Apparel and General Merchandise
\item Entertainment and Recreation
\item Grocery
\item Health Care
\item Resturants and Hotels
\item Transportation
}

###By Consumer Zip Code Income. Transactions are linked to zip codes where the consumer lives and zip codes are classified into income categories based on measurements of median household income and population provided by the American Community Survey (2014 - 2018).
\itemize{
\item High Income (median household income greater than $78,000 per year)
\item Middle Income (median household income between $46,000 per year and $78,000 per year)
\item Low Income (median household income less than $46,000 per year)
\item spend_all: Seasonally adjusted credit/debit card spending relative to January 4-31 2020 in all merchant category codes (MCC), 7 day moving average.
\item spend_acf: Seasonally adjusted credit/debit card spending relative to January 4-31 2020 in accomodation and food service (ACF) MCCs, 7 day moving average, 7 day moving average.
\item spend_aer: Seasonally adjusted credit/debit card spending relative to January 4-31 2020 in arts, entertainment, and recreation (AER) MCCs, 7 day moving average.
\item spend_apg: Seasonally adjusted credit/debit card spending relative to January 4-31 2020 in general merchandise stores (GEN) and apparel and accessories (AAP) MCCs, 7 day moving average.
\item spend_grf: Seasonally adjusted credit/debit card spending relative to January 4-31 2020 in grocery and food store (GRF) MCCs, 7 day moving average.
\item spend_hcs: Seasonally adjusted credit/debit card spending relative to January 4-31 2020 in health care and social assistance (HCS) MCCs, 7 day moving average.
\item spend_tws: Seasonally adjusted credit/debit card spending relative to January 4-31 2020 in transportation and warehousing (TWS) MCCs, 7 day moving average.
\item spend_all_inchigh: Seasonally adjusted credit/debit card spending by consumers living in ZIP codes with high (top quartile) median income, relative to January 4-31 2020 in all merchant category codes (MCC), 7 day moving average.
\item spend_all_incmiddle: Seasonally adjusted credit/debit card spending by consumers living in ZIP codes with middle (middle two quartiles) median income, relative to January 4-31 2020 in all merchant category codes (MCC), 7 day moving average.
\item spend_all_inclow: Seasonally adjusted credit/debit card spending by consumers living in ZIP codes with low (bottom quartiles) median income, relative to January 4-31 2020 in all merchant category codes (MCC), 7 day moving average.
\item spend_all_q2: Seasonally adjusted credit/debit card spending by consumers living in ZIP codes in the second quartile (i.e. second lowest) of median incomes, relative to January 4-31 2020 in all merchant category codes (MCC), 7 day moving average.
\item spend_all_q3: Seasonally adjusted credit/debit card spending by consumers living in ZIP codes in the third quartile (i.e. second highest) of median incomes, relative to January 4-31 2020 in all merchant category codes (MCC), 7 day moving average.
}
}

}
}
\note{
The raw data contains discontinuous breaks caused by entry or exit of credit card providers from the sample. In order to reliably identify and correct these breaks, we require at least 3 weeks of data. The most recent 3 weeks of data are therefore marked 'provisional' and are subject to non-negligible changes as new data is posted. For breaks found prior to the last 3 weeks, we correct for it using a method outlined in the paper. Otherwise we substitute the national mean for more recent breaks while we gather enough data to implement the corrections outlined in the paper. Additionally, at the county-level when are there more than one structural breaks the data is too noisy to correct for these breaks and counties with multiple breaks are dropped from the sample. Lastly, Affinity Solutions suppresses any cut of the data with fewer than five transactions. For more details refer to the accompanying paper.
}
\references{
\itemize{
\item \url{https://opportunityinsights.org/wp-content/uploads/2020/05/tracker_paper.pdf}
}
}
\seealso{
Other data-import: 
\code{\link{acaps_government_measures_data}()},
\code{\link{acaps_secondary_impact_data}()},
\code{\link{apple_mobility_data}()},
\code{\link{beoutbreakprepared_data}()},
\code{\link{cci_us_vaccine_data}()},
\code{\link{cdc_aggregated_projections}()},
\code{\link{cdc_excess_deaths}()},
\code{\link{cdc_social_vulnerability_index}()},
\code{\link{coronadatascraper_data}()},
\code{\link{coronanet_government_response_data}()},
\code{\link{cov_glue_lineage_data}()},
\code{\link{cov_glue_newick_data}()},
\code{\link{cov_glue_snp_lineage}()},
\code{\link{covidtracker_data}()},
\code{\link{descartes_mobility_data}()},
\code{\link{ecdc_data}()},
\code{\link{econ_tracker_unemp_data}},
\code{\link{economist_excess_deaths}()},
\code{\link{financial_times_excess_deaths}()},
\code{\link{google_mobility_data}()},
\code{\link{government_policy_timeline}()},
\code{\link{jhu_data}()},
\code{\link{jhu_us_data}()},
\code{\link{kff_icu_beds}()},
\code{\link{nytimes_county_data}()},
\code{\link{oecd_unemployment_data}()},
\code{\link{owid_data}()},
\code{\link{param_estimates_published}()},
\code{\link{test_and_trace_data}()},
\code{\link{us_county_geo_details}()},
\code{\link{us_county_health_rankings}()},
\code{\link{us_healthcare_capacity}()},
\code{\link{us_hospital_details}()},
\code{\link{us_state_distancing_policy}()},
\code{\link{usa_facts_data}()},
\code{\link{who_cases}()}
}
\author{
Sean Davis \href{mailto:seandavi@gmail.com}{seandavi@gmail.com}
}
\concept{data-import}
\concept{econonics}
