% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econ_tracker_employment.R
\name{econ_tracker_employment}
\alias{econ_tracker_employment}
\alias{econ_tracker_employment_city_data}
\alias{econ_tracker_employment_county_data}
\alias{econ_tracker_employment_state_data}
\alias{econ_tracker_employment_national_data}
\title{United States Employment Data from Econ Tracker and Opportunity Insights}
\usage{
econ_tracker_employment_city_data()

econ_tracker_employment_county_data()

econ_tracker_employment_state_data()

econ_tracker_employment_national_data()
}
\value{
a data.table with geographic details and a variable column with the following values representing the reported value:
\itemize{
\item emp_combined: Employment level for all workers.
\item emp_combined_inclow: Employment level for workers in the bottom quartile of the income distribution (incomes approximately under $27,000).
\item emp_combined_incmiddle: Employment level for workers in the middle two quartiles of the income distribution (incomes approximately $27,000 to $60,000).
\item emp_combined_inchigh: Employment level for workers in the top quartile of the income distribution (incomes approximately over $60,000).
\item emp_combined_ss40: Employment level for workers in trade, transportation and utilities (NAICS supersector 40).
\item emp_combined_ss60: Employment level for workers in professional and business services (NAICS supersector 60).
\item emp_combined_ss65: Employment level for workers in education and health services (NAICS supersector 65).
\item emp_combined_ss70: Employment level for workers in leisure and hospitality (NAICS supersector 70).
}
}
\description{
Number of active employees, aggregating information from multiple data providers. This series is based on firm-level payroll data from Paychex and Intuit, worker-level data on employment and earnings from Earnin, and firm-level timesheet data from Kronos.
}
\details{
Data Sources: Paychex, Intuit, Earnin, Kronos

Breakdowns are by:
\subsection{Wages}{
\itemize{
\item High Income (wage greater than $60,000 per year)
\item Middle Income (wage between $27,000 per year and $60,000 per year)
\item Low Income (wage less than $27,000 per year)
}
}

\subsection{Industry by \href{https://www.bls.gov/sae/additional-resources/naics-supersectors-for-ces-program.htm}{NAICS supersector}.}{
\itemize{
\item Professional and Business Services
\item Education and Health Services
\item Retail and Transportation
\item Leisure and Hospitality
}
}
}
\note{
For low income workers, the change in employment is calculated using Paychex and Earnin data. For medium and high income workers, the change in employment is calculated using Paychex and Intuit data.

In order to provide closer to real time data, we forecast the most recent employment measures beyond those available in the combined Earnin, Intuit, and Paychex dataset alone. To do so, we leverage two sources of higher frequency data: Kronos timestamp data and the Paychex weekly pay cycle sample. Using this higher frequency data we forecast more recent changes in employment using a distributed lag model, constructed by regressing a given week’s employment measure on the corresponding week’s Kronos measure, as well as its current and 3 previous lagged weeks’ Paychex weekly pay cycle measure. For more details, please refer to the appendix of the accompanying paper
}
\examples{

res = econ_tracker_employment_city_data()
res

res = econ_tracker_employment_county_data()
res


res = econ_tracker_employment_state_data()
res

res = econ_tracker_employment_national_data()
res

ggplot(res, aes(x=date,y=value,color=variable,lty=variable_type)) + 
    geom_line() + 
    ggtitle('Proportion of workforce employed in the United States', 
            subtitle='Broken down by industry and income level, normalized to January 2020') +
    ylab('Proportion of workforce') 

}
\seealso{
Other data-import: 
\code{\link{acaps_government_measures_data}()},
\code{\link{acaps_secondary_impact_data}()},
\code{\link{apple_mobility_data}()},
\code{\link{beoutbreakprepared_data}()},
\code{\link{cci_us_vaccine_data}()},
\code{\link{cdc_aggregated_projections}()},
\code{\link{cdc_excess_deaths}()},
\code{\link{cdc_social_vulnerability_index}()},
\code{\link{coronadatascraper_data}()},
\code{\link{coronanet_government_response_data}()},
\code{\link{cov_glue_lineage_data}()},
\code{\link{cov_glue_newick_data}()},
\code{\link{cov_glue_snp_lineage}()},
\code{\link{covidtracker_data}()},
\code{\link{descartes_mobility_data}()},
\code{\link{ecdc_data}()},
\code{\link{econ_tracker_consumer_spending}},
\code{\link{econ_tracker_unemp_data}},
\code{\link{economist_excess_deaths}()},
\code{\link{financial_times_excess_deaths}()},
\code{\link{google_mobility_data}()},
\code{\link{government_policy_timeline}()},
\code{\link{jhu_data}()},
\code{\link{jhu_us_data}()},
\code{\link{kff_icu_beds}()},
\code{\link{nytimes_county_data}()},
\code{\link{oecd_unemployment_data}()},
\code{\link{owid_data}()},
\code{\link{param_estimates_published}()},
\code{\link{test_and_trace_data}()},
\code{\link{us_county_geo_details}()},
\code{\link{us_county_health_rankings}()},
\code{\link{us_healthcare_capacity}()},
\code{\link{us_hospital_details}()},
\code{\link{us_state_distancing_policy}()},
\code{\link{usa_facts_data}()},
\code{\link{who_cases}()}

Other economics: 
\code{\link{acaps_secondary_impact_data}()},
\code{\link{econ_tracker_consumer_spending}},
\code{\link{econ_tracker_unemp_data}},
\code{\link{us_county_health_rankings}()}
}
\author{
Sean Davis \href{mailto:seandavi@gmail.com}{seandavi@gmail.com}
}
\concept{data-import}
\concept{economics}
