% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econ_tracker_data.R
\name{econ_tracker_unemp_data}
\alias{econ_tracker_unemp_data}
\alias{econ_tracker_unemp_city_data}
\alias{econ_tracker_unemp_county_data}
\alias{econ_tracker_unemp_state_data}
\alias{econ_tracker_unemp_national_data}
\title{Econ Tracker Unemployment Insurance Claims}
\source{
\url{https://github.com/OpportunityInsights/EconomicTracker}
}
\usage{
econ_tracker_unemp_city_data()

econ_tracker_unemp_county_data()

econ_tracker_unemp_state_data()

econ_tracker_unemp_national_data()
}
\description{
Unemployment insurance claims data from the \href{https://oui.doleta.gov/unemploy/DataDashboard.asp}{Department of Labor} (national and state-level) and numerous individual state agencies (county-level).
}
\details{
All datasets contain the following columns in addition to the location details. Dates specify the last day of the week ending
for the reported data.
\itemize{
\item initclaims_rate_regular: Number of initial claims per 100 people in the 2019 labor force, Regular UI only
\item initclaims_count_regular: Count of initial claims, Regular UI only
\item initclaims_rate_pua: Number of initial claims per 100 people in the 2019 labor force, PUA (Pandemic Unemployment Assistance) only
\item initclaims_count_pua: Count of initial claims, PUA (Pandemic Unemployment Assistance) only
\item initclaims_rate_combined: Number of initial claims per 100 people in the 2019 labor force, combining Regular and PUA claims
\item initclaims_count_combined: Count of initial claims, combining Regular and PUA claims
\item contclaims_rate_regular: Number of continued claims per 100 people in the 2019 labor force, Regular UI only
\item contclaims_count_regular: Count of continued claims, Regular UI only
\item contclaims_rate_pua: Number of continued claims per 100 people in the 2019 labor force, PUA (Pandemic Unemployment Assistance) only
\item contclaims_count_pua: Count of continued claims, PUA (Pandemic Unemployment Assistance) only
\item contclaims_rate_peuc: Number of continued claims per 100 people in the 2019 labor force, PEUC (Pandemic Emergency Unemployment Compensation) only
\item contclaims_count_peuc: Count of continued claims, PEUC (Pandemic Emergency Unemployment Compensation) only
\item contclaims_rate_combined: Number of continued claims per 100 people in the 2019 labor force, combining Regular, PUA and PEUC claims
\item contclaims_count_combined: Count of continued claims, combining Regular, PUA and PEUC claims
}
}
\examples{

# City Level data

res = econ_tracker_unemp_city_data()
head(res)



# County Level data

res = econ_tracker_unemp_county_data()
head(res)



# State Level data

res = econ_tracker_unemp_state_data()
head(res)



# National Level data

res = econ_tracker_unemp_national_data()
head(res)

}
\references{
\itemize{
\item \url{https://github.com/OpportunityInsights/EconomicTracker}
\item \url{https://tracktherecovery.org/}
}
}
\seealso{
Other data-import: 
\code{\link{acaps_government_measures_data}()},
\code{\link{acaps_secondary_impact_data}()},
\code{\link{apple_mobility_data}()},
\code{\link{beoutbreakprepared_data}()},
\code{\link{cci_us_vaccine_data}()},
\code{\link{cdc_aggregated_projections}()},
\code{\link{cdc_excess_deaths}()},
\code{\link{cdc_social_vulnerability_index}()},
\code{\link{coronadatascraper_data}()},
\code{\link{coronanet_government_response_data}()},
\code{\link{cov_glue_lineage_data}()},
\code{\link{cov_glue_newick_data}()},
\code{\link{cov_glue_snp_lineage}()},
\code{\link{covidtracker_data}()},
\code{\link{descartes_mobility_data}()},
\code{\link{ecdc_data}()},
\code{\link{econ_tracker_consumer_spending}},
\code{\link{economist_excess_deaths}()},
\code{\link{financial_times_excess_deaths}()},
\code{\link{google_mobility_data}()},
\code{\link{government_policy_timeline}()},
\code{\link{jhu_data}()},
\code{\link{jhu_us_data}()},
\code{\link{kff_icu_beds}()},
\code{\link{nytimes_county_data}()},
\code{\link{oecd_unemployment_data}()},
\code{\link{owid_data}()},
\code{\link{param_estimates_published}()},
\code{\link{test_and_trace_data}()},
\code{\link{us_county_geo_details}()},
\code{\link{us_county_health_rankings}()},
\code{\link{us_healthcare_capacity}()},
\code{\link{us_hospital_details}()},
\code{\link{us_state_distancing_policy}()},
\code{\link{usa_facts_data}()},
\code{\link{who_cases}()}

Other economics: 
\code{\link{acaps_secondary_impact_data}()},
\code{\link{econ_tracker_consumer_spending}},
\code{\link{us_county_health_rankings}()}
}
\concept{data-import}
\concept{economics}
