% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/economist_excess_deaths.R
\name{economist_excess_deaths}
\alias{economist_excess_deaths}
\title{Excess deaths during covid for US and selected countries}
\source{
\itemize{
\item \url{https://github.com/TheEconomist/covid-19-excess-deaths-tracker}
}
}
\usage{
economist_excess_deaths()
}
\description{
Return the data behind The Economist’s tracker for \href{https://www.economist.com/graphic-detail/2020/04/16/tracking-covid-19-excess-deaths-across-countries}{covid-19 excess deaths}
(which is free to read)
}
\examples{
res = economist_excess_deaths()
colnames(res)
dplyr::glimpse(res)
res
unique(res$country)

}
\references{
\itemize{
\item \url{https://www.economist.com/graphic-detail/2020/04/16/tracking-covid-19-excess-deaths-across-countries}
}
}
\seealso{
Other data-import: 
\code{\link{acaps_government_measures_data}()},
\code{\link{apple_mobility_data}()},
\code{\link{beoutbreakprepared_data}()},
\code{\link{cdc_aggregated_projections}()},
\code{\link{cdc_excess_deaths}()},
\code{\link{cdc_social_vulnerability_index}()},
\code{\link{coronadatascraper_data}()},
\code{\link{coronanet_government_response_data}()},
\code{\link{cov_glue_lineage_data}()},
\code{\link{cov_glue_newick_data}()},
\code{\link{cov_glue_snp_lineage}()},
\code{\link{covidtracker_data}()},
\code{\link{descartes_mobility_data}()},
\code{\link{ecdc_data}()},
\code{\link{financial_times_excess_deaths}()},
\code{\link{google_mobility_data}()},
\code{\link{government_policy_timeline}()},
\code{\link{jhu_data}()},
\code{\link{jhu_us_data}()},
\code{\link{kff_icu_beds}()},
\code{\link{nytimes_county_data}()},
\code{\link{oecd_unemployment_data}()},
\code{\link{owid_data}()},
\code{\link{param_estimates_published}()},
\code{\link{test_and_trace_data}()},
\code{\link{us_county_geo_details}()},
\code{\link{us_county_health_rankings}()},
\code{\link{us_healthcare_capacity}()},
\code{\link{us_hospital_details}()},
\code{\link{us_state_distancing_policy}()},
\code{\link{usa_facts_data}()},
\code{\link{who_cases}()},
\code{\link{yougov_behavioral_data}()}

Other excess-deaths: 
\code{\link{cdc_excess_deaths}()},
\code{\link{financial_times_excess_deaths}()}
}
\author{
Sean Davis \href{mailto:seandavi@gmail.com}{seandavi@gmail.com}
}
\concept{data-import}
\concept{excess-deaths}
