% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econ_tracker_data.R
\name{econtracker}
\alias{econtracker}
\alias{econ_tracker_county_geo}
\alias{econ_tracker_city_geo}
\title{Econ tracker information}
\source{
\itemize{
\item \url{https://github.com/OpportunityInsights/EconomicTracker}
}
}
\usage{
econ_tracker_county_geo()

econ_tracker_city_geo()
}
\value{
data.frame
}
\description{
Econ tracker information
}
\references{
\itemize{
\item \url{https://tracktherecovery.org/}
}
}
