% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app1.R
\name{eesi}
\alias{eesi}
\title{demonstrative app for EpiEstim R\link{t} sensitivity analysis}
\usage{
eesi(src)
}
\arguments{
\item{src}{intended to be the result of sars2pack::nytimes_state_data(), no checking is done.}
}
\description{
demonstrative app for EpiEstim R\link{t} sensitivity analysis
}
\note{
This app is currently incomplete.  Additions needed include 1) accommodation of
standard deviation of parametric serial interval distribution, 2) additional
approaches to serial interval modeling, 3) enhancement of displays to present actual
calendar dates, 4) an approach to supporting comparative displays, particularly allowing
overlay of different R\link{t} fits for different serial interval modeling assumptions.
}
\examples{
\dontrun{
 nytdat = nytimes_state_data()
 eesi(nytdat)
}
}
