% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jhu_to_excel.R
\name{enriched_jhu_data}
\alias{enriched_jhu_data}
\title{Get enriched jhu dataset enriched with country-level metadata}
\usage{
enriched_jhu_data(cols_to_remove = .cols_to_remove, warn = FALSE)
}
\arguments{
\item{cols_to_remove}{a character vector of column names from
\code{\link[=country_metadata]{country_metadata()}} to remove.}

\item{warn}{logical(1) defaults to FALSE, sent to \code{countrycode::countrycode} to
determine whether failed matches should be reported.}
}
\value{
A list of three \code{data.frames} named \code{deaths}, \code{confirmed},
and \code{recovered}.
}
\description{
The purpose of this function is to enrich the JHU dataset
with additional country-level metadata and then get the
data into a form that can be easily written to excel or used
as a "wide-format" tabular dataset.
}
\examples{
res = enriched_jhu_data()
colnames(res)
library(dplyr)
head(res)
glimpse(res)

}
