% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyEG.R
\name{est_R0}
\alias{est_R0}
\title{helper for easyEG, reach in to estimate structure}
\usage{
est_R0(x)
}
\arguments{
\item{x}{result of R0::estimate.R}
}
\description{
helper for easyEG, reach in to estimate structure
}
