% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_Rt.R
\name{est_Rt}
\alias{est_Rt}
\title{produce an MCMC-based EpiEstim R\link{t} estimate on the 29-day series}
\usage{
est_Rt(aout)
}
\arguments{
\item{aout}{instance of 'Arima_sars2pack' S3 class}
}
\description{
produce an MCMC-based EpiEstim R\link{t} estimate on the 29-day series
}
\note{
Simply uses the tsfull and dates29 components to call EpiEstim::estimate_R
}
\examples{
nyd = nytimes_state_data()
ny = Arima_by_state(nyd, state.in="New York")
ee = est_Rt(ny)
od = options()$digits
options(digits=3)
tail(ee$R)
options(digits=od)
}
