% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{estimate_R}
\alias{estimate_R}
\title{estimate R0}
\usage{
estimate_R(x, ...)
}
\arguments{
\item{x}{a data.frame with incidences}

\item{...}{passed along to dispatched function}
}
\value{
a data.frame
}
\description{
estimate R0
}
