% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eu-data-cache-data.R
\name{eu_data_cache_data}
\alias{eu_data_cache_data}
\alias{eu_data_cache_datasets}
\title{Country, region, and province level COVID-19 time series for Europe}
\source{
\itemize{
\item \url{https://covid19-eu-data-cache.now.sh/}
\item \url{https://github.com/covid19-eu-zh/covid19-eu-data}
}
}
\usage{
eu_data_cache_data(datasets = eu_data_cache_datasets(include_ecdc = FALSE))

eu_data_cache_datasets(include_ecdc = TRUE)
}
\description{
covid19-eu-data is a dataset repository for COVID-19/SARS-CoV-2
cases in Europe. They pull data from official government websites
regularly using the open-source scripts inside the repository.
}
\section{Functions}{
\itemize{
\item \code{eu_data_cache_datasets}: 
}}

\note{
Some oddities of the dataset are captured on the original
site at
\url{https://github.com/covid19-eu-zh/covid19-eu-data#notes}
}
