% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morefield.R
\name{extract_country_data}
\alias{extract_country_data}
\title{extract a sub data.table (row) with specific CountryRegion}
\usage{
extract_country_data(name, data)
}
\arguments{
\item{name}{character(1) row selection}

\item{data}{a data.table with field \code{CountryRegion}}
}
\description{
extract a sub data.table (row) with specific CountryRegion
}
