% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morefield.R
\name{extract_ProvinceState_data}
\alias{extract_ProvinceState_data}
\title{extract a sub data.table (row) with specific ProvinceState}
\usage{
extract_ProvinceState_data(df, name)
}
\arguments{
\item{df}{a data.frame with field \code{ProvinceState}}

\item{name}{character(1) row selection}
}
\value{
a filtered data.frame
}
\description{
extract a sub data.table (row) with specific ProvinceState
}
\examples{
jhu_data() \%>\% extract_ProvinceState_data("Maryland")

}
