% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fin_times_excess_deaths.R
\name{financial_times_excess_deaths}
\alias{financial_times_excess_deaths}
\title{Excess deaths tracking by Financial Times}
\source{
\url{https://github.com/Financial-Times/coronavirus-excess-mortality-data}
}
\usage{
financial_times_excess_deaths()
}
\description{
From the Financial Times: "'Excess mortality' refers to the difference between deaths from all causes during the pandemic and the historic seasonal average. For many of the jurisdictions shown here, this figure is higher than the official Covid-19 fatalities that are published by national governments each day. While not all of these deaths are necessarily attributable to the disease, it does leave a number of unexplained deaths that suggests that the official figures of deaths attributed may significant undercounts of the pandemic's impact."
}
\details{
From the Financial Times github repository:

The data contains excess mortality data for the period covering the
2020 Covid-19 pandemic. The data has been gathered from national,
regional or municipal agencies that collect death registrations and
publish official mortality statistics. These original data were
reshaped into a standardised format by Financial Times journalists
to allow cross-national comparisons, and have been used to inform
the FT’s reporting on the pandemic.  The repository contains the
excess mortality data for all known jurisdictions which publish
all-cause mortality data meeting the following criteria:
\itemize{
\item daily, weekly or monthly level of granularity
\item includes equivalent historical data for at least one full year before 2020, and preferably at least five years (2015-2019)
includes data up to at least April 1, 2020
}

Most countries publish mortality data with a longer periodicity
(typically quarterly or even annually), a longer publication lag
time, or both. This sort of data is not suitable for ongoing
analysis during an epidemic and is therefore not included here.
}
\section{License}{

\url{https://creativecommons.org/licenses/by/4.0/}
}

\examples{
res = financial_times_excess_deaths()

head(res)
dplyr::glimpse(res)

}
\references{
Coronavirus tracked: the latest figures as countries start to reopenFree to read, \url{https://www.ft.com/content/a26fbf7e-48f8-11ea-aeb3-955839e06441}
}
\seealso{
Other data-import: 
\code{\link{apple_mobility_data}()},
\code{\link{beoutbreakprepared_data}()},
\code{\link{cdc_aggregated_projections}()},
\code{\link{cdc_excess_deaths}()},
\code{\link{cdc_social_vulnerability_index}()},
\code{\link{coronadatascraper_data}()},
\code{\link{coronanet_government_response_data}()},
\code{\link{cov_glue_lineage_data}()},
\code{\link{cov_glue_newick_data}()},
\code{\link{cov_glue_snp_lineage}()},
\code{\link{covidtracker_data}()},
\code{\link{descartes_mobility_data}()},
\code{\link{ecdc_data}()},
\code{\link{economist_excess_deaths}()},
\code{\link{eu_data_cache_data}()},
\code{\link{google_mobility_data}()},
\code{\link{government_policy_timeline}()},
\code{\link{healthdata_mobility_data}()},
\code{\link{healthdata_projections_data}()},
\code{\link{healthdata_testing_data}()},
\code{\link{jhu_data}()},
\code{\link{jhu_us_data}()},
\code{\link{kff_icu_beds}()},
\code{\link{nytimes_county_data}()},
\code{\link{oecd_unemployment_data}()},
\code{\link{owid_data}()},
\code{\link{param_estimates_published}()},
\code{\link{test_and_trace_data}()},
\code{\link{us_county_geo_details}()},
\code{\link{us_county_health_rankings}()},
\code{\link{us_healthcare_capacity}()},
\code{\link{us_hospital_details}()},
\code{\link{us_state_distancing_policy}()},
\code{\link{usa_facts_data}()},
\code{\link{who_cases}()}

Other excess-deaths: 
\code{\link{cdc_excess_deaths}()},
\code{\link{economist_excess_deaths}()}
}
\author{
Sean Davis \href{mailto:seandavi@gmail.com}{seandavi@gmail.com}
}
\concept{data-import}
\concept{excess-deaths}
