% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejhu.R
\name{form_incident_events}
\alias{form_incident_events}
\title{transform cumulative data to daily incidence}
\usage{
form_incident_events(cum)
}
\arguments{
\item{cum}{cumulative events instance}
}
\description{
transform cumulative data to daily incidence
}
\examples{
suppressWarnings({dat = enriched_jhu_data()})
cusa = cumulative_events_ejhu(dat, eventtype="confirmed",
 alpha3="USA")
cusa
cusa = trim_from(cusa, "2020-03-01")
iusa = form_incident_events(cusa)
plot(iusa)
gt = R0::generation.time("gamma", c(3.5, 4.8))
est1 = R0::estimate.R( iusa$count, GT=gt,
   t=iusa$dates, begin=iusa$dates[1], end=iusa$dates[length(iusa$dates)],
   methods="EG")
est1
plot2(est1)
plotfit2(est1)
}
