% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{get_series}
\alias{get_series}
\title{supersimple series extraction}
\usage{
get_series(
  province = "",
  country,
  dataset = try(fetch_JHU_Data(as.data.frame = TRUE))
)
}
\arguments{
\item{province}{character(1) must be found in dataset ProvinceState field, "" is typical for data aggregated only at country level, and that is the default}

\item{country}{character(1) must be found in CountryRegion field}

\item{dataset}{data.frame as returned by fetch_JHU_Data}
}
\description{
supersimple series extraction
}
