% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_mobility.R
\name{google_mobility_data}
\alias{google_mobility_data}
\title{Google mobility dataset}
\source{
\itemize{
\item \url{https://www.google.com/covid19/mobility/}
}
}
\usage{
google_mobility_data(accept_terms = TRUE)
}
\arguments{
\item{accept_terms}{default is TRUE, but please make sure that you
and any downstream users are aware that they are accepting
Google's terms of service. See the note below.}
}
\description{
From the Google website: These Community Mobility Reports aim to
provide insights into what has changed in response to policies
aimed at combating COVID-19. The reports chart movement trends over
time by geography, across different categories of places such as
retail and recreation, groceries and pharmacies, parks, transit
stations, workplaces, and residential.
}
\details{
From Google:

What’s a Community Mobility Report?

Each Community Mobility Report is broken down by location and
displays the percent change from baseline in visits to places like
grocery stores and parks.
}
\note{
Using this function implies that you agree to the Google terms of service:
\url{https://policies.google.com/terms}
}
\section{Privacy}{


Google provides this explanation with regard to preserving privacy:

The Community Mobility Reports were developed to be helpful while
adhering to our stringent privacy protocols and protecting people’s
privacy. No personally identifiable information, such as an
individual’s location, contacts or movement, will be made available
at any point.

Insights in these reports are created with aggregated, anonymized
sets of data from users who have turned on the Location History
setting, which is off by default. People who have Location History
turned on can choose to turn it off at any time from their Google
Account and can always delete Location History data directly from
their Timeline.

We also use the same world-class anonymization technology used in
our products every day to keep your activity data private and
secure. This includes differential privacy, which adds artificial
noise to our datasets, enabling us to generate insights while
preventing the identification of any individual person.
}

\examples{

res = google_mobility_data()
colnames(res)
head(res)
dplyr::glimpse(res)

# analyze the "admin levels" available for each
# country:

admin_by_country = res \%>\%
    dplyr::group_by(iso2c) \%>\%
    dplyr::filter(date == max(date))
admin_by_country = table(admin_by_country$admin_level, admin_by_country$iso2c)/length(unique(res$places_category))
admin_by_country

# Italy mobility over time
# Note day-of week effect in plot

library(ggplot2)

res \%>\%
## Italy, whole country(admin == 0)
    filter(iso2c == "IT" & admin_level==0) \%>\%
    ggplot(aes(x=date,y=percent_change_from_baseline,color = places_category)) +
    geom_line() +
    ggtitle('Google mobility metric for Italy') +
    theme(legend.position='bottom')


}
\seealso{
Other data-import: 
\code{\link{apple_mobility_data}()},
\code{\link{beoutbreakprepared_data}()},
\code{\link{cdc_aggregated_projections}()},
\code{\link{cdc_excess_deaths}()},
\code{\link{cov_glue_lineage_data}()},
\code{\link{cov_glue_newick_data}()},
\code{\link{covidtracker_data}()},
\code{\link{descartes_mobility_data}()},
\code{\link{ecdc_data}()},
\code{\link{eu_data_cache_data}()},
\code{\link{government_policy_timeline}()},
\code{\link{healthdata_projections_data}()},
\code{\link{jhu_data}()},
\code{\link{jhu_us_data}()},
\code{\link{kff_icu_beds}()},
\code{\link{nytimes_county_data}()},
\code{\link{param_estimates_published}()},
\code{\link{us_county_geo_details}()},
\code{\link{us_healthcare_capacity}()},
\code{\link{us_state_distancing_policy}()},
\code{\link{usa_facts_data}()}

Other mobility: 
\code{\link{apple_mobility_data}()},
\code{\link{descartes_mobility_data}()}
}
\author{
Sean Davis \href{mailto:seandavi@gmail.com}{seandavi@gmail.com}
}
\concept{data-import}
\concept{mobility}
