% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxford_govt_policy_timeline.R
\name{government_policy_timeline}
\alias{government_policy_timeline}
\title{Longitudinal tracking of government interventions to combat COVID-19}
\source{
\itemize{
\item \url{https://www.bsg.ox.ac.uk/research/research-projects/oxford-covid-19-government-response-tracker}
}
}
\usage{
government_policy_timeline()
}
\description{
Governments are taking a wide range of measures in response to the COVID-19 outbreak. The Oxford COVID-19 Government Response Tracker (OxCGRT) aims to track and compare government responses to the coronavirus outbreak worldwide rigorously and consistently.
}
\details{
Systematic information on which governments have taken which measures, and when, can help decision-makers and citizens understand the robustness of governmental responses in a consistent way, aiding efforts to fight the pandemic. The OxCGRT systematically collects information on several different common policy responses governments have taken, scores the stringency of such measures, and aggregates these scores into a common Stringency Index.

Data is collected from public sources by a team of dozens of Oxford University students and staff from every part of the world.

OxCGRT collects publicly available information on 11 indicators of government response (S1-S11). The first seven indicators (S1-S7) take policies such as school closures, travel bans, etc. are recorded on an ordinal scale; the remainder (S8-S11) are financial indicators such as fiscal or monetary measures. For a full description of the data and how it is collected, see this working paper.

Recommended citation for database: Hale, Thomas and Samuel Webster (2020). Oxford COVID-19 Government Response Tracker. Data use policy: Creative Commons Attribution CC BY standard.
}
\examples{
library(dplyr)
res = government_policy_timeline()
head(res)
colnames(res)
dplyr::glimpse(res)
summary(res)

}
\seealso{
Other data-import: 
\code{\link{apple_mobility_data}()},
\code{\link{beoutbreakprepared_data}()},
\code{\link{cdc_aggregated_projections}()},
\code{\link{covidtracker_data}()},
\code{\link{descartes_mobility_data}()},
\code{\link{ecdc_data}()},
\code{\link{eu_data_cache_data}()},
\code{\link{google_mobility_data}()},
\code{\link{healthdata_projections_data}()},
\code{\link{jhu_data}()},
\code{\link{jhu_us_data}()},
\code{\link{kff_icu_beds}()},
\code{\link{nytimes_county_data}()},
\code{\link{param_estimates_published}()},
\code{\link{us_county_geo_details}()},
\code{\link{us_healthcare_capacity}()},
\code{\link{us_state_distancing_policy}()},
\code{\link{usa_facts_data}()}

Other non-pharmaceutical-interventions: 
\code{\link{us_state_distancing_policy}()}
}
\author{
Sean Davis \href{mailto:seandavi@gmail.com}{seandavi@gmail.com}
}
\concept{data-import}
\concept{non-pharmaceutical-interventions}
