% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arima.R
\name{grid_2d}
\alias{grid_2d}
\title{a two-dimensional grid of integers}
\usage{
grid_2d(n = 5)
}
\arguments{
\item{n}{maximum value, will expand.grid(0:n,0:n) with colnames arorder, maorder}
}
\description{
a two-dimensional grid of integers
}
