% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthdata_projections_data.R
\name{healthdata_projections_data}
\alias{healthdata_projections_data}
\title{healthdata.org covid19 morbidity and hospitalization estimates}
\source{
\itemize{
\item \url{https://ihmecovid19storage.blob.core.windows.net/latest/ihme-covid19.zip}
}
}
\usage{
healthdata_projections_data()
}
\value{
data.frame with columns for:
\itemize{
\item location_name
\item date
\item metric: such as all beds needed
\item mean, upper, lower: mean, upper
confidence interval, lower confidence interval
}
}
\description{
These are time-series data that forecaset the COVID-19 impact
on hospital bed-days, ICU-days, ventilator-days, and deaths by
US state and selected additional countries or regions. The data
are those used here \url{https://covid19.healthdata.org/projections}.
}
\examples{
res = healthdata_projections_data()
colnames(res)
res[sample(1:nrow(res),6),]

}
\references{
\itemize{
\item \url{https://www.medrxiv.org/content/10.1101/2020.03.27.20043752v1}
}
}
\seealso{
\itemize{
\item \url{http://www.healthdata.org/covid}
\item \url{http://www.healthdata.org/covid/updates}
\item \url{http://www.healthdata.org/covid/faqs}
}

Other data-import: 
\code{\link{apple_mobility_data}()},
\code{\link{beoutbreakprepared_data}()},
\code{\link{covidtracker_data}()},
\code{\link{descartes_mobility_data}()},
\code{\link{jhu_data}()},
\code{\link{jhu_us_data}()},
\code{\link{kff_icu_beds}()},
\code{\link{nytimes_county_data}()},
\code{\link{us_county_geo_details}()},
\code{\link{us_healthcare_capacity}()},
\code{\link{us_state_distancing_policy}()},
\code{\link{usa_facts_data}()}
}
\author{
Sean Davis \href{mailto:seandavi@gmail.com}{seandavi@gmail.com}
}
\concept{data-import}
