% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthdata_projections_data.R
\name{healthdata_testing_data}
\alias{healthdata_testing_data}
\title{healthdata.org covid19 testing observations and predictions}
\source{
\itemize{
\item \url{https://ihmecovid19storage.blob.core.windows.net/latest/ihme-covid19.zip}
}
}
\usage{
healthdata_testing_data()
}
\description{
These are time-series data that forecaset testing quantities for
US state and selected additional countries or regions. The data
are those used here \url{https://covid19.healthdata.org/projections}.
}
\examples{
res = healthdata_mobility_data()
res
summary(res)
dplyr::glimpse(res)

}
\references{
\itemize{
\item \url{https://www.medrxiv.org/content/10.1101/2020.03.27.20043752v1}
}
}
\seealso{
\itemize{
\item \url{http://www.healthdata.org/covid}
\item \url{http://www.healthdata.org/covid/updates}
\item \url{http://www.healthdata.org/covid/faqs}
}

Other data-import: 
\code{\link{apple_mobility_data}()},
\code{\link{beoutbreakprepared_data}()},
\code{\link{cdc_aggregated_projections}()},
\code{\link{cdc_excess_deaths}()},
\code{\link{cdc_social_vulnerability_index}()},
\code{\link{coronanet_government_response_data}()},
\code{\link{cov_glue_lineage_data}()},
\code{\link{cov_glue_newick_data}()},
\code{\link{cov_glue_snp_lineage}()},
\code{\link{covidtracker_data}()},
\code{\link{descartes_mobility_data}()},
\code{\link{ecdc_data}()},
\code{\link{economist_excess_deaths}()},
\code{\link{eu_data_cache_data}()},
\code{\link{google_mobility_data}()},
\code{\link{government_policy_timeline}()},
\code{\link{healthdata_mobility_data}()},
\code{\link{healthdata_projections_data}()},
\code{\link{jhu_data}()},
\code{\link{jhu_us_data}()},
\code{\link{kff_icu_beds}()},
\code{\link{nytimes_county_data}()},
\code{\link{oecd_unemployment_data}()},
\code{\link{owid_data}()},
\code{\link{param_estimates_published}()},
\code{\link{us_county_geo_details}()},
\code{\link{us_county_health_rankings}()},
\code{\link{us_healthcare_capacity}()},
\code{\link{us_hospital_details}()},
\code{\link{us_state_distancing_policy}()},
\code{\link{usa_facts_data}()}
}
\author{
Sean Davis \href{mailto:seandavi@gmail.com}{seandavi@gmail.com}
}
\concept{case-series}
\concept{data-import}
