% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morefield.R
\name{jhu_data}
\alias{jhu_data}
\title{retrieve time series dataset from CSSEGIS}
\usage{
jhu_data()
}
\value{
an object of class \code{s2p_long_df} that inherits from tbl_df
}
\description{
This always does a web call. The assumption here is that
the user is online and that github is up and that URLs are
stable.
}
\note{
Uses https://raw.githubusercontent.com/CSSEGISandData/... as data
source, then modifies column names and munges to long form table.
}
\examples{
res = jhu_data()
colnames(res)
head(res)

}
