% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jhu_to_excel.R
\name{jhu_data_to_excel}
\alias{jhu_data_to_excel}
\title{Create +/- write excel format of enriched JHU global data}
\usage{
jhu_data_to_excel(
  dat = jhu_data(),
  excel_filename = NA,
  cols_to_remove = c("callingCodes", "altSpellings", "latlng", "demonym", "timezones",
    "nativeName", "currencies", "languages", "translations", "flag", "regionalBlocs"),
  ...
)
}
\arguments{
\item{dat}{a data.frame-like object with at least column
\code{CountryRegion} that will be joined with country data.}

\item{excel_filename}{character(1) filename to which to save excel
file}

\item{cols_to_remove}{a character vector of column names from
\code{\link[=country_metadata]{country_metadata()}} to remove.}
}
\value{
A list of three \code{data.frames} named \code{deaths}, \code{confirmed},
and \code{recovered}.
}
\description{
The purpose of this function is to enrich the JHU dataset
with additional country-level metadata and then get the
data into a form that can be easily written to excel or used
as a "wide-format" tabular dataset.
}
