% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jhu_data.R
\name{jhu_us_data}
\alias{jhu_us_data}
\title{US counties COVID-19 data from \href{https://github.com/CSSEGISandData/COVID-19/}{JHU CSSEGIS}}
\source{
\itemize{
\item \url{https://github.com/CSSEGISandData/COVID-19/tree/master/csse_covid_19_data/csse_covid_19_time_series}
}
}
\usage{
jhu_us_data()
}
\value{
A tidy \code{data.frame} (actually, a \code{tbl_df}) with columns:
\itemize{
\item UID:  Universal Identifier
\item iso2:  ISO 3166-1 alpha-2 code
\item iso3:  ISO 3166-1 alpha-3 code
\item code3
\item fips:  Federal Information Processing Standard Publication code
\item county:  County
\item state:  Province or state.
\item country:  US
\item Lat:  Latitude
\item Long:  Longitude
\item Combined_Key:  Comma-separated combination of columns \code{Admin2}, \code{ProvinceState}, and \code{CountryRegion}
\item date:  Date
\item count:  The cumulative count of cases for a given geographic area.
\item subset:  either \code{confirmed} or \code{deaths}
}
}
\description{
This function access and munges the cumulative time series of confirmed,
and deaths from the US data in the repository for the 2019 Novel Coronavirus Visual
Dashboard operated by the Johns Hopkins University Center for
Systems Science and Engineering (JHU CSSE). Also, Supported by ESRI
Living Atlas Team and the Johns Hopkins University Applied Physics
Lab (JHU APL).
}
\details{
Data are updated daily by JHU. Each call to this function redownloads the data
from github. No data cleansing is performed. Data are downloaded and then munged
into long-form tidy \code{data.frame}.
}
\note{
Uses https://raw.githubusercontent.com/CSSEGISandData/... as data
source, then modifies column names and munges to long form table.

\itemize{
\item Although numbers are meant to be cumulative, there are instances where a day's count might
be less than the prior day due to a reclassification of a case. These are not currently corrected
in the source data
}
}
\examples{
res = jhu_data()
colnames(res)
head(res)
dplyr::glimpse(res)

table(res$state)


}
\seealso{
Other data-import: 
\code{\link{apple_mobility_data}()},
\code{\link{beoutbreakprepared_data}()},
\code{\link{cdc_aggregated_projections}()},
\code{\link{cdc_excess_deaths}()},
\code{\link{coronanet_government_response_data}()},
\code{\link{cov_glue_lineage_data}()},
\code{\link{cov_glue_newick_data}()},
\code{\link{covidtracker_data}()},
\code{\link{descartes_mobility_data}()},
\code{\link{ecdc_data}()},
\code{\link{eu_data_cache_data}()},
\code{\link{google_mobility_data}()},
\code{\link{government_policy_timeline}()},
\code{\link{healthdata_projections_data}()},
\code{\link{jhu_data}()},
\code{\link{kff_icu_beds}()},
\code{\link{nytimes_county_data}()},
\code{\link{param_estimates_published}()},
\code{\link{us_county_geo_details}()},
\code{\link{us_county_health_rankings}()},
\code{\link{us_healthcare_capacity}()},
\code{\link{us_hospital_details}()},
\code{\link{us_state_distancing_policy}()},
\code{\link{usa_facts_data}()}
}
\concept{data-import}
