% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kff_icu_beds.R
\name{kff_icu_beds}
\alias{kff_icu_beds}
\title{Kaiser Family Foundation US County-level ICU bed data}
\source{
\itemize{
\item \url{https://khn.org/news/as-coronavirus-spreads-widely-millions-of-older-americans-live-in-counties-with-no-icu-beds/}
\item \url{https://khn.org/wp-content/uploads/sites/2/2020/03/KHN-ICU-bed-county-analysis_2.zip}
}
}
\usage{
kff_icu_beds()
}
\description{
Kaiser Family Foundation US County-level ICU bed data
}
\examples{
res = kff_icu_beds()
colnames(res)
head(res)

}
\seealso{
Other data-import: 
\code{\link{beoutbreakprepared_data}()},
\code{\link{covidtracker_data}()},
\code{\link{decartes_mobility_data}()},
\code{\link{healthdata_projections_data}()},
\code{\link{jhu_data}()},
\code{\link{jhu_us_data}()},
\code{\link{nytimes_county_data}()},
\code{\link{us_county_geo_details}()},
\code{\link{us_healthcare_capacity}()},
\code{\link{usa_facts_data}()}
}
\concept{data-import}
