% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_utils.R
\name{ls_github}
\alias{ls_github}
\title{do an ls of a github repo}
\usage{
ls_github(repository, path = "", ref = "master")
}
\arguments{
\item{repository}{character(1) formatted as 'user/repo_name' from Github}

\item{path}{character(1) The path in the repo to \code{ls} on}

\item{ref}{character(1) The github reference (tag, sha, branch) to
run the ls on. Default is master.}
}
\value{
A \code{tibble} with columns including:
\itemize{
\item name
\item path
\item size
\item download_url
\item type (\code{file} or \code{dir})
}
}
\description{
Specify repo, path, and github reference (master, sha, etc.)
and get back an \code{ls} of that directory, including download links.
}
\examples{
ls_github('TheEconomist/covid-19-excess-deaths-tracker',
          'output-data/excess-deaths')

}
