% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cumulative_events.R
\name{make_cumulative_events}
\alias{make_cumulative_events}
\title{generic cumulative_events generator}
\usage{
make_cumulative_events(cases, dates, eventtype, alpha3, ProvinceStateName)
}
\arguments{
\item{cases}{numeric vector}

\item{dates}{vector}

\item{eventtype}{character(1) should be "confirmed" or "deaths"}

\item{alpha3}{character(1) country code not checked}

\item{ProvinceStateName}{character(1)}
}
\description{
generic cumulative_events generator
}
\examples{
ec = jhu_data()
ber = ec[which(ec$nuts_3 == "Bergamo"),]
mber = make_cumulative_events( cases = ber$cases, dates = ber$date, 
  eventtype = "confirmed", alpha3 = "ITA", ProvinceStateName = "Bergamo")
plot(mber)
}
