% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mar19df}
\alias{mar19df}
\title{march 19 result of CSSEGISandData/COVID-19/ time series data retrieval as data.frame}
\format{
data.frame
}
\usage{
mar19df
}
\description{
march 19 result of CSSEGISandData/COVID-19/ time series data retrieval as data.frame
}
\examples{
mar19df[1:3, 1:8]
}
\keyword{datasets}
