% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mar19dt}
\alias{mar19dt}
\title{march 19 result of CSSEGISandData/COVID-19/ time series data retrieval as data.table}
\format{
data.frame
}
\usage{
mar19dt
}
\description{
march 19 result of CSSEGISandData/COVID-19/ time series data retrieval as data.table
}
\examples{
mar19dt[1:3, 1:8]
}
\keyword{datasets}
