% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arima.R
\name{min_bic}
\alias{min_bic}
\title{minimize BIC for ARIMA models with default differencing order 1 over choices of MA and AR order}
\usage{
min_bic(
  src,
  fullusa = FALSE,
  state.in = "New York",
  parms = grid_2d(5),
  max_date = NULL,
  lookback_days = 29,
  ...
)
}
\arguments{
\item{src}{tibble with cumulative incidence like output of nytimes_state_data()}

\item{fullusa}{logical(1) if TRUE, use Arima_nation (src should be enriched_jhu_data())}

\item{state.in}{character(1) state name}

\item{parms}{two-column data frame with proposed AR order in column 1 and MA order in column 2}

\item{max_date}{character(1) or lubridate date from which to look back}

\item{lookback_days}{numeric(1)}

\item{\dots}{passed to \code{Arima_by_state}}
}
\description{
minimize BIC for ARIMA models with default differencing order 1 over choices of MA and AR order
}
\examples{
nyd = nytimes_state_data()
mb = min_bic(nyd, state.in="New York")
names(mb)
}
