% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{min_bic_2020_05_20}
\alias{min_bic_2020_05_20}
\title{a statewide collection of best AR/MA model orders}
\format{
S3 instance
}
\usage{
min_bic_2020_05_20
}
\description{
a statewide collection of best AR/MA model orders
}
\keyword{datasets}
