% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arima.R
\name{min_bic_all_states}
\alias{min_bic_all_states}
\title{organize search over contiguous states for optimal statewise AR/MA using BIC}
\usage{
min_bic_all_states(src, ...)
}
\arguments{
\item{src}{tibble with cumulative incidence like output of nytimes_state_data()}

\item{\dots}{passed to \code{min_bic}}
}
\value{
instance of S3 class \code{min_bic_all_states}
}
\description{
organize search over contiguous states for optimal statewise AR/MA using BIC
}
