% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextstrain.R
\name{nextstrain_case_data}
\alias{nextstrain_case_data}
\title{Nextstrain/GISAID metadata table (raw input)}
\source{
\itemize{
\item \url{https://nextstrain.org/}
\item \url{https://raw.githubusercontent.com/nextstrain/ncov/master/data/metadata.tsv}
}
}
\usage{
nextstrain_case_data()
}
\description{
This function gathers the raw input data associated with
\url{https://nextstrain.org/} for COVID-19 as gathered from GISAID.
}
\details{
Nextstrain is an open-source project to harness the scientific and
public health potential of pathogen genome data. They provide a
continually-updated view of publicly available data alongside
powerful analytic and visualization tools for use by the
community. Their goal is to aid epidemiological understanding and
improve outbreak response.
}
\section{Nextstrain overview}{

From the Nextstrain website:

In the course of an infection and over an epidemic, pathogens
naturally accumulate random mutations to their genomes. This is an
inevitable consequence of error-prone genome replication. Since
different genomes typically pick up different mutations, mutations
can be used as a marker of transmission in which closely related
genomes indicate closely related infections. By reconstructing a
phylogeny we can learn about important epidemiological phenomena
such as spatial spread, introduction timings and epidemic growth
rate.

However, if pathogen genome sequences are going to inform public
health interventions, then analyses have to be rapidly conducted
and results widely disseminated. Current scientific publishing
practices hinder the rapid dissemination of epidemiologically
relevant results. We thought an open online system that implements
robust bioinformatic pipelines to synthesize data from across
research groups has the best capacity to make epidemiologically
actionable inferences.

Workflow for Nextstrain processing:
\itemize{
\item \url{https://github.com/nextstrain/ncov/raw/master/docs/images/basic_snakemake_build.png}
}
}

\examples{
res = nextstrain_case_data()
colnames(res)
head(res)
library(dplyr)
dplyr::glimpse(res)

}
\references{
James Hadfield, Colin Megill, Sidney M Bell, John Huddleston,
Barney Potter, Charlton Callender, Pavel Sagulenko, Trevor Bedford,
Richard A Neher, Nextstrain: real-time tracking of pathogen
evolution, Bioinformatics, Volume 34, Issue 23, 01 December 2018,
Pages 4121–4123,
\url{https://doi.org/10.1093/bioinformatics/bty407}
}
\seealso{
Other data-import: 
\code{\link{apple_mobility_data}()},
\code{\link{beoutbreakprepared_data}()},
\code{\link{cdc_aggregated_projections}()},
\code{\link{cdc_excess_deaths}()},
\code{\link{covidtracker_data}()},
\code{\link{descartes_mobility_data}()},
\code{\link{ecdc_data}()},
\code{\link{eu_data_cache_data}()},
\code{\link{google_mobility_data}()},
\code{\link{government_policy_timeline}()},
\code{\link{healthdata_projections_data}()},
\code{\link{jhu_data}()},
\code{\link{jhu_us_data}()},
\code{\link{kff_icu_beds}()},
\code{\link{nytimes_county_data}()},
\code{\link{param_estimates_published}()},
\code{\link{us_county_geo_details}()},
\code{\link{us_healthcare_capacity}()},
\code{\link{us_state_distancing_policy}()},
\code{\link{usa_facts_data}()}

Other individual-cases: 
\code{\link{beoutbreakprepared_data}()}
}
\concept{data-import}
\concept{genomics}
\concept{individual-cases}
