% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nytimes_data.R
\name{nytimes_county_data}
\alias{nytimes_county_data}
\alias{nytimes_state_data}
\title{United States county and state level data from the New York Times}
\source{
\itemize{
\item \url{https://raw.githubusercontent.com/nytimes/covid-19-data/master/us-counties.csv}
\item \url{https://raw.githubusercontent.com/nytimes/covid-19-data/master/us-states.csv}
}
}
\usage{
nytimes_county_data()

nytimes_state_data()
}
\value{
a five-column (for states) or six-column (for counties) tibble
\itemize{
\item date: observation date
\item county: county (or sometimes city), see \url{https://github.com/nytimes/covid-19-data/} for details
\item state: present in both the state and county data. Note that simply aggregating by state will
sometimes overcount when working with the \code{county} level data. See \url{https://github.com/nytimes/covid-19-data/} for details.
\item fips: The Federal Information Processing Standard Publication 6-4 (FIPS 6-4)
was a five-digit Federal Information Processing Standards code which uniquely
identified counties and county equivalents in the United States, certain U.S.
possessions, and certain freely associated states.
\item count: number of cases (cumulative)
\item subset: \code{deaths} or \code{confirmed}
}
}
\description{
Each row of data reports cumulative counts based on best reporting
up to the moment published. The NYT states that it may revise
earlier entries in the data when they receive new information.
}
\details{
From the NYTimes github README:

The New York Times is releasing a series of data files with
cumulative counts of coronavirus cases in the United States, at the
state and county level, over time. We are compiling this time
series data from state and local governments and health departments
in an attempt to provide a complete record of the ongoing outbreak.

Since late January, The Times has tracked cases of coronavirus in
real time as they were identified after testing. Because of the
widespread shortage of testing, however, the data is necessarily
limited in the picture it presents of the outbreak.

We have used this data to power our maps and reporting tracking the
outbreak, and it is now being made available to the public in
response to requests from researchers, scientists and government
officials who would like access to the data to better understand
the outbreak.

The data begins with the first reported coronavirus case in
Washington State on Jan. 21, 2020. We will publish regular updates
to the data in this repository.
}
\note{
See \url{https://github.com/nytimes/covid-19-data#geographic-exceptions}. Also,
This dataset contains county data with "holes" in reporting, dates with no
reported results. Also, records with "Unknown" county are removed since
these records appear to NOT be cumulative data, but incidence data?
}
\section{Licensing}{


The NYTimes provides a license that states:

In general, we are making this data publicly available for broad,
noncommercial public use including by medical and public health
researchers, policymakers, analysts and local news media.

If you use this data, you must attribute it to “The New York Times”
in any publication. If you would like a more expanded description
of the data, you could say “Data from The New York Times, based on
reports from state and local health agencies.”

If you use it in an online presentation, we would appreciate it if
you would link to our U.S. tracking page at
\url{https://www.nytimes.com/interactive/2020/us/coronavirus-us-cases.html}.
}

\examples{
# state data
res = nytimes_state_data()
colnames(res)
dplyr::glimpse(res)

# county data
res = nytimes_county_data()
colnames(res)
dplyr::glimpse(res)

 
}
\seealso{
jhu_data(), usa_facts_data()

Other data-import: 
\code{\link{apple_mobility_data}()},
\code{\link{beoutbreakprepared_data}()},
\code{\link{cdc_aggregated_projections}()},
\code{\link{cdc_excess_deaths}()},
\code{\link{cov_glue_lineage_data}()},
\code{\link{cov_glue_newick_data}()},
\code{\link{covidtracker_data}()},
\code{\link{descartes_mobility_data}()},
\code{\link{ecdc_data}()},
\code{\link{eu_data_cache_data}()},
\code{\link{google_mobility_data}()},
\code{\link{government_policy_timeline}()},
\code{\link{healthdata_projections_data}()},
\code{\link{jhu_data}()},
\code{\link{jhu_us_data}()},
\code{\link{kff_icu_beds}()},
\code{\link{param_estimates_published}()},
\code{\link{us_county_geo_details}()},
\code{\link{us_healthcare_capacity}()},
\code{\link{us_state_distancing_policy}()},
\code{\link{usa_facts_data}()}

Other case-tracking: 
\code{\link{align_to_baseline}()},
\code{\link{beoutbreakprepared_data}()},
\code{\link{bulk_estimate_Rt}()},
\code{\link{combined_us_cases_data}()},
\code{\link{covidtracker_data}()},
\code{\link{ecdc_data}()},
\code{\link{estimate_Rt}()},
\code{\link{eu_data_cache_data}()},
\code{\link{jhu_data}()},
\code{\link{plot_epicurve}()},
\code{\link{usa_facts_data}()}
}
\author{
Sean Davis \href{mailto:seandavi@gmail.com}{seandavi@gmail.com}
}
\concept{case-tracking}
\concept{data-import}
