% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owid_data.R
\name{owid_data}
\alias{owid_data}
\title{Our world in data worldwide cases and tests}
\source{
\itemize{
\item \url{https://ourworldindata.org/coronavirus}
}
}
\usage{
owid_data()
}
\value{
a data.frame
}
\description{
The OurWoldInData dataset includes country-level testing,
deaths, and confirmed cases for most countries in the world.
}
\examples{
res = owid_data()
colnames(res)

head(res)

dplyr::glimpse(res)
  
}
\references{
Max Roser, Hannah Ritchie, Esteban Ortiz-Ospina and Joe Hasell (2020) -
"Coronavirus Pandemic (COVID-19)". Published online at OurWorldInData.org.
Retrieved from: 'https://ourworldindata.org/coronavirus'
}
\seealso{
Other case-tracking: 
\code{\link{align_to_baseline}()},
\code{\link{beoutbreakprepared_data}()},
\code{\link{bulk_estimate_Rt}()},
\code{\link{combined_us_cases_data}()},
\code{\link{coronadatascraper_data}()},
\code{\link{covidtracker_data}()},
\code{\link{ecdc_data}()},
\code{\link{estimate_Rt}()},
\code{\link{jhu_data}()},
\code{\link{nytimes_county_data}()},
\code{\link{plot_epicurve}()},
\code{\link{test_and_trace_data}()},
\code{\link{usa_facts_data}()},
\code{\link{who_cases}()}

Other data-import: 
\code{\link{acaps_government_measures_data}()},
\code{\link{acaps_secondary_impact_data}()},
\code{\link{apple_mobility_data}()},
\code{\link{beoutbreakprepared_data}()},
\code{\link{cdc_aggregated_projections}()},
\code{\link{cdc_excess_deaths}()},
\code{\link{cdc_social_vulnerability_index}()},
\code{\link{coronadatascraper_data}()},
\code{\link{coronanet_government_response_data}()},
\code{\link{cov_glue_lineage_data}()},
\code{\link{cov_glue_newick_data}()},
\code{\link{cov_glue_snp_lineage}()},
\code{\link{covidtracker_data}()},
\code{\link{descartes_mobility_data}()},
\code{\link{ecdc_data}()},
\code{\link{economist_excess_deaths}()},
\code{\link{financial_times_excess_deaths}()},
\code{\link{google_mobility_data}()},
\code{\link{government_policy_timeline}()},
\code{\link{jhu_data}()},
\code{\link{jhu_us_data}()},
\code{\link{kff_icu_beds}()},
\code{\link{nytimes_county_data}()},
\code{\link{oecd_unemployment_data}()},
\code{\link{param_estimates_published}()},
\code{\link{test_and_trace_data}()},
\code{\link{us_county_geo_details}()},
\code{\link{us_county_health_rankings}()},
\code{\link{us_healthcare_capacity}()},
\code{\link{us_hospital_details}()},
\code{\link{us_state_distancing_policy}()},
\code{\link{usa_facts_data}()},
\code{\link{who_cases}()}
}
\author{
Sean Davis \href{mailto:seandavi@gmail.com}{seandavi@gmail.com}
}
\concept{case-tracking}
\concept{data-import}
