% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_estimates_published.R
\name{param_estimates_published}
\alias{param_estimates_published}
\title{Published parameter estimates from the literature}
\source{
\itemize{
\item \url{https://github.com/midas-network/COVID-19/blob/master/parameter_estimates/2019_novel_coronavirus/estimates.csv}
}
}
\usage{
param_estimates_published()
}
\description{
Published estimates of epidemiological characteristics from both
peer-reviewed and non-peer-reviewed resources, encoded by community
members and approved by authors. For complete information on each
estimate, see the actual dataset items.
}
\examples{
library(dplyr)
params = param_estimates_published()
colnames(params)
head(params)
dplyr::glimpse(params)

summary(params)

if(require(datatable) & interactive()) {
    DT(params)
}

}
\references{
\itemize{
\item \url{https://midasnetwork.us/covid-19/}
}
}
\seealso{
Other data-import: 
\code{\link{apple_mobility_data}()},
\code{\link{beoutbreakprepared_data}()},
\code{\link{cdc_aggregated_projections}()},
\code{\link{covidtracker_data}()},
\code{\link{descartes_mobility_data}()},
\code{\link{ecdc_data}()},
\code{\link{eu_data_cache_data}()},
\code{\link{google_mobility_data}()},
\code{\link{government_policy_timeline}()},
\code{\link{healthdata_projections_data}()},
\code{\link{jhu_data}()},
\code{\link{jhu_us_data}()},
\code{\link{kff_icu_beds}()},
\code{\link{nytimes_county_data}()},
\code{\link{us_county_geo_details}()},
\code{\link{us_healthcare_capacity}()},
\code{\link{us_state_distancing_policy}()},
\code{\link{usa_facts_data}()}
}
\concept{data-import}
