% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejhu.R
\name{plot.covid_events}
\alias{plot.covid_events}
\title{plot for covid_events}
\usage{
\method{plot}{covid_events}(x, main = NULL, ylab = NULL, xlab = NULL, ...)
}
\arguments{
\item{x}{covid_events instance}

\item{main}{defaults to NULL, can be character(1)}

\item{ylab}{defaults to NULL, can be character(1)}

\item{xlab}{defaults to NULL, can be character(1)}

\item{\dots}{not used}
}
\description{
plot for covid_events
}
