% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{plot2.R0.sR}
\alias{plot2.R0.sR}
\title{version of R0:::plot.R0.sR that does not use dev.new}
\usage{
\method{plot2}{R0.sR}(x, xscale = "w", TD.split = FALSE, ...)
}
\arguments{
\item{x}{instance of R0.sR}

\item{xscale}{character(1) scale to be used on x axis (d = daily, w=weekly, f=fortnightly, m=monthly)}

\item{TD.split}{logical(1) "to force the display of both R(t) and the epidemic curve in the same window for TD method"}

\item{\dots}{passed to base::plot}
}
\value{
none, used for side effects
}
\description{
version of R0:::plot.R0.sR that does not use dev.new
}
