% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{plot2}
\alias{plot2}
\title{generic for simpler plotter for R0.sR}
\usage{
plot2(x, xscale = "w", TD.split = FALSE, ...)
}
\arguments{
\item{x}{R0.sR instance}

\item{xscale}{character code for x axis scale}

\item{TD.split}{logical}

\item{\dots}{passed to base::plot}
}
\description{
generic for simpler plotter for R0.sR
}
