% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{plotfit2.R0.sR}
\alias{plotfit2.R0.sR}
\title{version of R0:::plotfit.R0.sR that does not use dev.new}
\usage{
\method{plotfit2}{R0.sR}(x, all = TRUE, xscale = "w", SB.dist = TRUE, ...)
}
\arguments{
\item{x}{instance of R0.sR}

\item{all}{logical(1)}

\item{xscale}{character(1) scale to be used on x axis (d = daily, w=weekly, f=fortnightly, m=monthly)}

\item{SB.dist}{logical(1)}

\item{\dots}{passed to base::plot}
}
\value{
none, used for side effects
}
\description{
version of R0:::plotfit.R0.sR that does not use dev.new
}
