% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{plotfit2}
\alias{plotfit2}
\title{generic for simpler plotter for plotfit for R0.sR}
\usage{
plotfit2(x, all = TRUE, xscale = "w", SB.dist = TRUE, ...)
}
\arguments{
\item{x}{R0.sR instance}

\item{all}{logical(1)}

\item{xscale}{character code for x axis scale}

\item{SB.dist}{logical(1)}

\item{\dots}{passed to base::plot}
}
\description{
generic for simpler plotter for plotfit for R0.sR
}
