% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morefield.R
\name{prepare_data}
\alias{prepare_data}
\title{prepare data using data.table idiom, unlist, trimLeading etc.}
\usage{
prepare_data(cdata)
}
\arguments{
\item{cdata}{presumably a data.table}
}
\description{
prepare data using data.table idiom, unlist, trimLeading etc.
}
