% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arima.R
\name{print.bic_seq}
\alias{print.bic_seq}
\title{report for optimal AR/MA search using BIC}
\usage{
\method{print}{bic_seq}(x, ...)
}
\arguments{
\item{x}{instance of bic_seq S3 class}

\item{\dots}{not used}
}
\description{
report for optimal AR/MA search using BIC
}
