% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejhu.R
\name{print.covid_events}
\alias{print.covid_events}
\title{print for a covid_events instance}
\usage{
\method{print}{covid_events}(x, ...)
}
\arguments{
\item{x}{covid_events instance}

\item{\dots}{not used}
}
\description{
print for a covid_events instance
}
