% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regapp.R
\name{regapp}
\alias{regapp}
\title{demonstration 'dynamic' regional map, shows some issues with legend control, needs optimization}
\usage{
regapp(nytc, cg, indate = "2020-03-15")
}
\description{
demonstration 'dynamic' regional map, shows some issues with legend control, needs optimization
}
\examples{
if (interactive()) {
 if (!exists("nyt_counties")) nyt_counties = nytimes_county_data()
 if (!exists("county_geom")) county_geom = tidycensus::county_laea
 regapp( nyt_counties, county_geom, indate="2020-03-15")
}
}
