% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{s2p_get_cache}
\alias{s2p_get_cache}
\title{get the BiocFileCache for sars2pack}
\usage{
s2p_get_cache(cache = rappdirs::user_cache_dir(appname = "sars2pack"))
}
\arguments{
\item{cache}{character(1), the path to the cache directory. See \code{\link[BiocFileCache:BiocFileCache-class]{BiocFileCache::BiocFileCache()}}.}
}
\value{
a \code{BiocFileCache} instance
}
\description{
Set up or get location of BiocFileCache. By default,
this uses the name BiocFileCache.
}
\keyword{internal}
