% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sars2gb_2020_05_14}
\alias{sars2gb_2020_05_14}
\title{genbank records including available sequence for SARS-Cov-2 accessions available 14 May 2020}
\format{
list
}
\usage{
sars2gb_2020_05_14
}
\description{
genbank records including available sequence for SARS-Cov-2 accessions available 14 May 2020
}
\keyword{datasets}
