% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sars2pack-package.R
\docType{package}
\name{sars2pack-package}
\alias{sars2pack-package}
\title{sars2pack: Easy access to and use of iconic COVID-19 data resource}
\description{
On January 30, 2020, the World Health Organization declared
coronavirus disease 2019 (COVID-19) a Public Health Emergency of
International concern (PHEIC) and within six weeks had
characterized the outbreak as a pandemic. Compared to the 2003-2004
severe acute respiratory syndrome (SARS) PHEIC, the COVID-19
pandemic is spreading more quickly and with a much higher death
toll. However, the current pandemic is occurring in a more digital
and interconnected world. Traditional public health organizations
as well as data-mature organizations not traditionally involved
directly in public health have rapidly developed digital disease
surveillance infrastructuree that provides nearly realtime epidemic
tracking data. These data resources have proven invaluable to
understanding disease spread, to drive non-pharmacologic
intervention (NPI), and, when combined with additional data
resources, to project impacts to communities and healthcare systems
around the world. Even as the urgency of the initial “hammer” of
the COVID-19 pandemic begins to abate, the need for timely, robust,
and granular datasets will inform business, policy, and even
personal decisions for months or even years to come.
}
\section{Purpose}{


The \code{sars2pack} R package aims to:
\enumerate{
\item Collect COVID-19 related public health and disease tracking
resourcesand provide principled approach date reuse and
reproducible computational research.
\item Provide a data science environment for researchers, media,
policy makers, and data scientists to collaborate while promoting
reproducible computational research best practices.
\item Capitalize on the large, existing multidisciplinary data science
workforce already familiar with the R programming environment.
\item Create opportunities for individuals not well-versed in data
science to learn and experiment with COVID-19 datasets.
\item Incorporate examplar workflows that leverage the extensive R
data science ecosystem to visualize, analyze, and integrate
COVID-19 data resources.
}
}

\section{Foo functions}{

The foo functions ...
}

\examples{

library(sars2pack)

# Show me all the available datasets
available_datasets()

example(jhu_data)

example(apple_mobility_data)

example(healthdata_projections_data)



}
