% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{si_cens_W}
\alias{si_cens_W}
\title{MCMC-based model for SI, Weibull family used for censored-data likelihood}
\format{
list with components \code{si_sample} (matrix) and `si_parametric_distr (character)
}
\usage{
si_cens_W
}
\description{
MCMC-based model for SI, Weibull family used for censored-data likelihood
}
\keyword{datasets}
