% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{si_pos_G}
\alias{si_pos_G}
\title{MCMC-based model for SI, Gamma family used for positive-data likelihood}
\format{
list with components \code{si_sample} (matrix) and `si_parametric_distr (character)
}
\usage{
si_pos_G
}
\description{
MCMC-based model for SI, Gamma family used for positive-data likelihood
}
\note{
The data reported at \href{https://wwwnc.cdc.gov/eid/article/26/6/20-0357_article}{CDC} on
468 transmission events were filtered to include only reports where contact
symptom onset was later than that reported for index case.
}
\keyword{datasets}
