% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{si_pos_L}
\alias{si_pos_L}
\title{MCMC-based model for SI, lognormal family used for positive-data likelihood}
\format{
list with components \code{si_sample} (matrix) and `si_parametric_distr (character)
}
\usage{
si_pos_L
}
\description{
MCMC-based model for SI, lognormal family used for positive-data likelihood
}
\keyword{datasets}
