% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_single_dataset.R
\name{test_data_accessor}
\alias{test_data_accessor}
\title{Check a dataset to see if it matches the expected output}
\usage{
test_data_accessor(accessor, ...)
}
\arguments{
\item{accessor}{character(1) the accessor function name of the dataset to check.}

\item{...}{passed to accessor directly}
}
\description{
Check a dataset to see if it matches the expected output
}
\examples{

test_data_accessor("nytimes_state_data")

}
\author{
Sean Davis \href{mailto:seandavi@gmail.com}{seandavi@gmail.com}
}
\keyword{internal}
