% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejhu.R
\name{trim_between.covid_events}
\alias{trim_between.covid_events}
\title{focus on segment of series delimited by two dates}
\usage{
\method{trim_between}{covid_events}(x, date1 = "2020-02-15", date2 = "2020-03-15")
}
\arguments{
\item{x}{covid_events instance}

\item{date1}{string representing date in yyyy-mm-dd form}

\item{date2}{string representing date in yyyy-mm-dd form}
}
\description{
focus on segment of series delimited by two dates
}
