% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejhu.R
\name{trim_between}
\alias{trim_between}
\title{trim part of series, generic}
\usage{
trim_between(x, date1, date2)
}
\arguments{
\item{x}{covid_events instance}

\item{date1}{string representing date in yyyy-mm-dd form}

\item{date2}{string representing date in yyyy-mm-dd form}
}
\description{
trim part of series, generic
}
