% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejhu.R
\name{trim_from.covid_events}
\alias{trim_from.covid_events}
\title{trim early part of series}
\usage{
\method{trim_from}{covid_events}(x, date = "2020-02-15")
}
\arguments{
\item{x}{covid_events instance}

\item{date}{string representing date in yyyy-mm-dd form}
}
\description{
trim early part of series
}
