% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejhu.R
\name{trim_from}
\alias{trim_from}
\title{trim early part of series, generic}
\usage{
trim_from(x, date)
}
\arguments{
\item{x}{covid_events instance}

\item{date}{string representing date in yyyy-mm-dd form}
}
\description{
trim early part of series, generic
}
