% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimLeading.R
\name{trim_leading_values}
\alias{trim_leading_values}
\title{trim leading repeats of a given value from a vector}
\usage{
trim_leading_values(x, value = 0)
}
\arguments{
\item{x}{vector of the same type as \code{value}}

\item{value}{value of which to remove leading repeats}
}
\value{
a vector trimmed of all leading values equal to \code{value}
}
\description{
trim leading repeats of a given value from a vector
}
\examples{
trim_leading_values(c(0,0,0,1,2,3))
trim_leading_values(c('a','a','d','e'), 'a')

}
\author{
Charles Morefield
}
