% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morefield.R
\name{trimLeading}
\alias{trimLeading}
\title{trim leading repeats of given value}
\usage{
trimLeading(x, value = 0)
}
\arguments{
\item{x}{vector of same type as \code{value}}

\item{value}{entity whose repeats are to be removed}
}
\description{
trim leading repeats of given value
}
\examples{
dat = c(0,0,3,0,4)
trimLeading(dat, value=0)
}
