% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsapp.R
\name{tsapp}
\alias{tsapp}
\title{shiny app for appraisal of simple time-series models for COVID-19 incidence trajectories}
\usage{
tsapp()
}
\description{
shiny app for appraisal of simple time-series models for COVID-19 incidence trajectories
}
