% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsapp.R
\name{tsapp_old}
\alias{tsapp_old}
\title{shiny app for appraisal of simple time-series models for COVID-19 incidence trajectories (old)}
\usage{
tsapp_old()
}
\description{
shiny app for appraisal of simple time-series models for COVID-19 incidence trajectories (old)
}
