% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/us_counties_gazetteer.R
\name{us_county_geo_details}
\alias{us_county_geo_details}
\title{United States county-level geographic details}
\source{
\url{https://raw.githubusercontent.com/josh-byster/fips_lat_long/master/counties.txt}
}
\usage{
us_county_geo_details()
}
\value{
a data frame with county names, FIPS codes, areas, and lat/long
}
\description{
United States county-level geographic details
}
\examples{
us_county_geo_details()

}
\seealso{
\url{https://www.census.gov/geographies/reference-files/time-series/geo/gazetteer-files.html}
}
