% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/us_county_health_rankings.R
\name{us_county_health_rankings}
\alias{us_county_health_rankings}
\title{United States county health rankings}
\usage{
us_county_health_rankings()
}
\description{
The County Health Rankings, a collaboration between the Robert Wood Johnson Foundation and the University of Wisconsin Population Health Institute, measure the health of nearly all counties in the nation and rank them within states. The Rankings are compiled using county-level measures from a variety of national and state data sources.
}
\details{
The County Health Rankings gather data from around the nation that is comparable between counties,
within states. For most of their measures, county data is also comparable across state lines. However, for
a few of their measures, some caution must be exercised when making comparisons between counties in
different states. See references below for details.
}
\examples{
ret = us_county_health_rankings()
head(ret)
colnames(ret)


}
\references{
\itemize{
\item \url{https://www.countyhealthrankings.org/sites/default/files/media/document/2020\%20Analytic\%20Documentation_0.pdf}
\item \url{https://www.countyhealthrankings.org/sites/default/files/media/document/2020\%20Comparability\%20across\%20states.pdf}
}
}
\seealso{
Other data-import: 
\code{\link{apple_mobility_data}()},
\code{\link{beoutbreakprepared_data}()},
\code{\link{cdc_aggregated_projections}()},
\code{\link{cdc_excess_deaths}()},
\code{\link{coronanet_government_response_data}()},
\code{\link{cov_glue_lineage_data}()},
\code{\link{cov_glue_newick_data}()},
\code{\link{covidtracker_data}()},
\code{\link{descartes_mobility_data}()},
\code{\link{ecdc_data}()},
\code{\link{eu_data_cache_data}()},
\code{\link{google_mobility_data}()},
\code{\link{government_policy_timeline}()},
\code{\link{healthdata_projections_data}()},
\code{\link{jhu_data}()},
\code{\link{jhu_us_data}()},
\code{\link{kff_icu_beds}()},
\code{\link{nytimes_county_data}()},
\code{\link{param_estimates_published}()},
\code{\link{us_county_geo_details}()},
\code{\link{us_healthcare_capacity}()},
\code{\link{us_hospital_details}()},
\code{\link{us_state_distancing_policy}()},
\code{\link{usa_facts_data}()}
}
\author{
Sean Davis \href{mailto:seandavi@gmail.com}{seandavi@gmail.com}
}
\concept{data-import}
