% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_care_map.R
\name{us_healthcare_capacity}
\alias{us_healthcare_capacity}
\title{United States healthcare system capacity by provider}
\source{
\itemize{
\item \url{https://github.com/covidcaremap/covid19-healthsystemcapacity}
\item \url{https://raw.githubusercontent.com/covidcaremap/covid19-healthsystemcapacity/master/data/published/us_healthcare_capacity-facility-CovidCareMap.csv}
}
}
\usage{
us_healthcare_capacity()
}
\description{
United States healthcare system capacity by provider
}
\details{
From the data providers:

Mapping existing and forecasted health system capacity gaps (beds,
staffing, ventilators, supplies) to care for surging numbers of
COVID19 patients (especially ICU-level care) at high spatiotemporal
resolution (by facility, daily, all USA to start).
}
\examples{
us_healthcare_capacity()

}
