% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/us_population_details.R
\name{us_population_details}
\alias{us_population_details}
\title{Retrieve United States Population data from the US Census Bureau}
\usage{
us_population_details(...)
}
\arguments{
\item{\dots}{Passed along to \code{\link[tidycensus]{get_estimates}}.}
}
\description{
Get data from the US Census Bureau Population Estimates APIs.
}
\examples{
res = us_population_details(geometry='state',year=2019)
head(res)

}
\author{
Sean Davis \href{mailto:seandavi@gmail.com}{seandavi@gmail.com}
}
