% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usa_facts_data.R
\name{usa_facts_data}
\alias{usa_facts_data}
\title{retrieve time series dataset from USA Facts}
\usage{
usa_facts_data()
}
\value{
an object of class \code{s2p_long_df} that inherits from tbl_df
}
\description{
This always does a web call. The assumption here is that
the user is online and that github is up and that URLs are
stable.
}
\note{
Uses https://usafacts.org/visualizations/coronavirus-covid-19-spread-map/ as data
source, then modifies column names and munges to long form table.
}
\examples{
res = usa_facts_data()
colnames(res)
head(res)

}
