% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world_population_data.R
\name{world_population_data}
\alias{world_population_data}
\title{World population details from United Nations}
\source{
\itemize{
\item \url{https://population.un.org}
}
}
\usage{
world_population_data()
}
\value{
a \code{data.frame}
}
\description{
World population details from United Nations
}
\examples{
res = world_population_data()

colnames(res)
dplyr::glimpse(res)
head(res)


}
\author{
Sean Davis \href{mailto:seandavi@gmail.com}{seandavi@gmail.com}
}
