% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_catalog.R
\name{available_datasets}
\alias{available_datasets}
\title{list and detail available sars2pack datasets}
\usage{
available_datasets()
}
\value{
A data.frame with columns:
\itemize{
\item name: the short name of the dataset
\item accessor: accessor function
}
}
\description{
list and detail available sars2pack datasets
}
\examples{
available_datasets()

}
