% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beoutbreakprepared.R
\name{beoutbreakprepared_data}
\alias{beoutbreakprepared_data}
\title{United States state and county level data from USAFacts}
\source{
https://github.com/beoutbreakprepared/nCoV2019
}
\usage{
beoutbreakprepared_data(quietly = TRUE)
}
\arguments{
\item{quietly}{logical(1) defaults to TRUE.  If FALSE, warnings generated during
parsing will be displayed.  These often relate to nonstandard date values that
occur idiosyncratically.}
}
\value{
tidy data.frame of content
}
\description{
United States state and county level data from USAFacts
}
\note{
This is individual level data, collected upstream of
this package from diverse sources.  Date values in particular
can be non-standard.  If you wish to introduce corrections for
poorly formated data, see the argument to \code{read.csv} in the source
code of this function and acquire the table manually.
}
\examples{
res = beoutbreakprepared_data()
colnames(res)
head(res)

}
\seealso{
Other data-import: 
\code{\link{covidtracker_data}()},
\code{\link{jhu_data}()},
\code{\link{kff_icu_beds}()},
\code{\link{nytimes_county_data}()},
\code{\link{us_county_geo_details}()},
\code{\link{us_healthcare_capacity}()},
\code{\link{usa_facts_data}()}
}
\concept{data-import}
