% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beoutbreakprepared.R
\name{beoutbreakprepared_data}
\alias{beoutbreakprepared_data}
\title{Individual-level data contributed from around the world}
\source{
https://github.com/beoutbreakprepared/nCoV2019
}
\usage{
beoutbreakprepared_data(quietly = TRUE)
}
\arguments{
\item{quietly}{logical(1) defaults to TRUE.  If FALSE, warnings generated during
parsing will be displayed.  These often relate to nonstandard date values that
occur idiosyncratically.}
}
\value{
tidy data.frame of content
}
\description{
Individual-level data contributed from around the world
}
\note{
This is individual level data, collected from diverse sources. Data may be messy
and we have made limited attempts at clean up.
}
\section{Citation}{

misc{kraemer2020epidemiological,
author =       {nCoV-2019 Data Working Group},
title =        {{Epidemiological Data from the nCoV-2019 Outbreak: Early
Descriptions from Publicly Available Data}},
howpublished = {Accessed on yyyy-mm-dd from
\url{http://virological.org/t/epidemiological-data-from-the-ncov-2019-outbreak-early-descriptions-from-publicly-available-data/337}},
year =         2020
}
}

\examples{
res = beoutbreakprepared_data()
colnames(res)
head(res)

}
\seealso{
Other data-import: 
\code{\link{covidtracker_data}()},
\code{\link{jhu_data}()},
\code{\link{jhu_us_data}()},
\code{\link{kff_icu_beds}()},
\code{\link{nytimes_county_data}()},
\code{\link{us_county_geo_details}()},
\code{\link{us_healthcare_capacity}()},
\code{\link{usa_facts_data}()}
}
\concept{data-import}
