% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_glue_lineage_data.R
\name{cov_glue_newick_data}
\alias{cov_glue_newick_data}
\title{COVID GLUE Newick tree}
\usage{
cov_glue_newick_data()
}
\description{
COVID GLUE Newick tree
}
\examples{
tr = cov_glue_newick_data()
plot(tr, type='r', cex=0.4)

}
\references{
\itemize{
\item \url{https://www.biorxiv.org/content/10.1101/2020.04.17.046086v1}
}
}
\seealso{
cov_glue_lineage_data(), cov_glue_snp_lineage()

Other lineage: 
\code{\link{cov_glue_lineage_data}()},
\code{\link{cov_glue_snp_lineage}()}

Other data-import: 
\code{\link{apple_mobility_data}()},
\code{\link{beoutbreakprepared_data}()},
\code{\link{cdc_aggregated_projections}()},
\code{\link{cdc_excess_deaths}()},
\code{\link{cdc_social_vulnerability_index}()},
\code{\link{coronanet_government_response_data}()},
\code{\link{cov_glue_lineage_data}()},
\code{\link{cov_glue_snp_lineage}()},
\code{\link{covidtracker_data}()},
\code{\link{descartes_mobility_data}()},
\code{\link{ecdc_data}()},
\code{\link{economist_excess_deaths}()},
\code{\link{eu_data_cache_data}()},
\code{\link{google_mobility_data}()},
\code{\link{government_policy_timeline}()},
\code{\link{healthdata_mobility_data}()},
\code{\link{healthdata_projections_data}()},
\code{\link{healthdata_testing_data}()},
\code{\link{jhu_data}()},
\code{\link{jhu_us_data}()},
\code{\link{kff_icu_beds}()},
\code{\link{nytimes_county_data}()},
\code{\link{oecd_unemployment_data}()},
\code{\link{owid_data}()},
\code{\link{param_estimates_published}()},
\code{\link{us_county_geo_details}()},
\code{\link{us_county_health_rankings}()},
\code{\link{us_healthcare_capacity}()},
\code{\link{us_hospital_details}()},
\code{\link{us_state_distancing_policy}()},
\code{\link{usa_facts_data}()}

Other individual-cases: 
\code{\link{beoutbreakprepared_data}()},
\code{\link{cov_glue_lineage_data}()}
}
\concept{data-import}
\concept{individual-cases}
\concept{lineage}
