% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidtracking_com.R
\name{covidtracker_data}
\alias{covidtracker_data}
\title{US State-level data including test results}
\usage{
covidtracker_data()
}
\value{
A tidy \code{tbl_df}
}
\description{
Get data from the COVID tracking project
\url{https://covidtracking.com/} including daily historical data on
testing results, hospitalizations, and deaths.
}
\details{
From the COVID Tracking Project website:

The COVID Tracking Project collects information from 50 US states,
the District of Columbia, and 5 other US territories to provide the
most comprehensive testing data we can collect for the novel
coronavirus, SARS-CoV-2. We attempt to report positive and negative
results, pending tests, and total people tested for each state or
district currently reporting that data.
}
\examples{
library(dplyr)
library(ggplot2)

res = covidtracker_data()
colnames(res)
dim(res)
dplyr::glimpse(res)

# Hospitalizations by day in Maryland
covidtracker_data() \%>\%
    dplyr::filter(state=='MD') \%>\%
    add_incidence_column(count_column='hospitalized') \%>\%
    ggplot(aes(x=date,y=inc)) + geom_smooth() +
    ylab("New Hospitalizations per day") +
    ggtitle('Hospitalizations in Maryland', subtitle = 'From covidtracker')


}
\seealso{
\itemize{
\item \url{https://covidtracking.com/about-tracker/}
\item \url{https://covidtracking.com/notes/}
}

Other data-import: 
\code{\link{apple_mobility_data}()},
\code{\link{beoutbreakprepared_data}()},
\code{\link{cdc_aggregated_projections}()},
\code{\link{cdc_excess_deaths}()},
\code{\link{coronanet_government_response_data}()},
\code{\link{cov_glue_lineage_data}()},
\code{\link{cov_glue_newick_data}()},
\code{\link{descartes_mobility_data}()},
\code{\link{ecdc_data}()},
\code{\link{eu_data_cache_data}()},
\code{\link{google_mobility_data}()},
\code{\link{government_policy_timeline}()},
\code{\link{healthdata_projections_data}()},
\code{\link{jhu_data}()},
\code{\link{jhu_us_data}()},
\code{\link{kff_icu_beds}()},
\code{\link{nytimes_county_data}()},
\code{\link{param_estimates_published}()},
\code{\link{us_county_geo_details}()},
\code{\link{us_county_health_rankings}()},
\code{\link{us_healthcare_capacity}()},
\code{\link{us_hospital_details}()},
\code{\link{us_state_distancing_policy}()},
\code{\link{usa_facts_data}()}

Other case-tracking: 
\code{\link{align_to_baseline}()},
\code{\link{beoutbreakprepared_data}()},
\code{\link{bulk_estimate_Rt}()},
\code{\link{combined_us_cases_data}()},
\code{\link{ecdc_data}()},
\code{\link{estimate_Rt}()},
\code{\link{eu_data_cache_data}()},
\code{\link{jhu_data}()},
\code{\link{nytimes_county_data}()},
\code{\link{owid_data}()},
\code{\link{plot_epicurve}()},
\code{\link{usa_facts_data}()}
}
\concept{case-tracking}
\concept{data-import}
