% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejhu.R
\name{cumulative_events_ejhu}
\alias{cumulative_events_ejhu}
\title{constructor for covid_events}
\usage{
cumulative_events_ejhu(
  src,
  eventtype = "confirmed",
  alpha3 = "USA",
  ProvinceStateName = NULL
)
}
\arguments{
\item{src}{as retrieved with enhanced_jhu_data}

\item{eventtype}{character(1) 'confirmed' or 'deaths'}

\item{alpha3}{character(1) code for country}

\item{ProvinceStateName}{character(1) for province, default to NULL; use NA for multiregion like GBR}
}
\description{
constructor for covid_events
}
\examples{
dat = enriched_jhu_data()
cchn = cumulative_events_ejhu(dat, alpha3="CHN", ProvinceStateName="Hubei")
ichn = form_incident_events(cchn)
plot(cchn$count[-1], ichn$count, log="xy", xlab="cumulative", ylab="new", 
 main="Hubei province") # like https://youtu.be/54XLXg4fYsc
}
