% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eu-data-cache-data.R
\name{eu_data_cache_data}
\alias{eu_data_cache_data}
\alias{eu_data_cache_datasets}
\title{Country, region, and province level COVID-19 time series for Europe}
\source{
\itemize{
\item \url{https://covid19-eu-data-cache.now.sh/}
\item \url{https://github.com/covid19-eu-zh/covid19-eu-data}
}
}
\usage{
eu_data_cache_data(datasets = eu_data_cache_datasets(include_ecdc = FALSE))

eu_data_cache_datasets(include_ecdc = TRUE)
}
\description{
covid19-eu-data is a dataset repository for COVID-19/SARS-CoV-2
cases in Europe. They pull data from official government websites
regularly using the open-source scripts inside the repository.
}
\section{Functions}{
\itemize{
\item \code{eu_data_cache_datasets}: 
}}

\note{
Some oddities of the dataset are captured on the original
site at
\url{https://github.com/covid19-eu-zh/covid19-eu-data#notes}
}
\examples{
eudata = eu_data_cache_data()
colnames(eudata)
dim(eudata)
head(eudata)
dplyr::glimpse(eudata)
summary(eudata)

# You may wish to look at only one or a few
# of the available datasets. Here is the list
eu_data_cache_datasets()


}
\seealso{
Other data-import: 
\code{\link{apple_mobility_data}()},
\code{\link{beoutbreakprepared_data}()},
\code{\link{covidtracker_data}()},
\code{\link{descartes_mobility_data}()},
\code{\link{ecdc_data}()},
\code{\link{google_mobility_data}()},
\code{\link{healthdata_projections_data}()},
\code{\link{jhu_data}()},
\code{\link{jhu_us_data}()},
\code{\link{kff_icu_beds}()},
\code{\link{nytimes_county_data}()},
\code{\link{us_county_geo_details}()},
\code{\link{us_healthcare_capacity}()},
\code{\link{us_state_distancing_policy}()},
\code{\link{usa_facts_data}()}
}
\concept{data-import}
