% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morefield.R
\name{extract_country_data}
\alias{extract_country_data}
\title{extract a set of rows with specific CountryRegion}
\usage{
extract_country_data(data, name)
}
\arguments{
\item{name}{character(1) row selection}

\item{df}{a data.frame with field \code{CountryRegion}}
}
\value{
a filtered data.frame
}
\description{
extract a set of rows with specific CountryRegion
}
