% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morefield.R
\name{extract_ProvinceState_data}
\alias{extract_ProvinceState_data}
\title{extract a sub data.table (row) with specific ProvinceState}
\usage{
extract_ProvinceState_data(name, data)
}
\arguments{
\item{name}{character(1) row selection}

\item{data}{a data.table with field `CountryRegion`}
}
\description{
extract a sub data.table (row) with specific ProvinceState
}
