% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integer_to_fips.R
\name{integer_to_fips}
\alias{integer_to_fips}
\title{convert an integer to fips}
\usage{
integer_to_fips(v)
}
\arguments{
\item{integer}{vector of length 1 or more}
}
\value{
character vector with each member a 0-padded representation of the original
integer value.
}
\description{
The Federal Information Processing Standards (fips) codes
are available as 5-digit, 0-padded numbers. To standardize
these codes for joins across datasets, one may need to convert
raw integers (non-0-padded) to 5-digit numeric strings. This
function does that.
}
\examples{
integer_to_fips(50)
class(integer_to_fips(50))
nchar(integer_to_fips(50))==5
integer_to_fips(sample(1:50000,5))

}
\references{
List of United States FIPS codes by county: \url{https://en.wikipedia.org/wiki/List_of_United_States_FIPS_codes_by_county}.
}
