% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jhu_data.R
\name{jhu_data}
\alias{jhu_data}
\title{Global COVID-19 data from \href{https://github.com/CSSEGISandData/COVID-19/}{JHU CSSEGIS}}
\source{
\itemize{
\item \url{https://github.com/CSSEGISandData/COVID-19/tree/master/csse_covid_19_data/csse_covid_19_time_series,mGT, method=c('EG','TD'))}
}
}
\usage{
jhu_data()
}
\value{
A tidy \code{data.frame} (actually, a \code{tbl_df}) with columns:
\itemize{
\item ProvinceState:  Province or state. \strong{Note}:
\item CountryRegion:  This is the main column for finding countries of interest
\item Lat:  Latitude
\item Long:  Longitude
\item date:  Date
\item count:  The cumulative count of cases for a given geographic area.
\item subset: either \code{confirmed}, \code{deaths}, or \code{recovered}.
}
}
\description{
This function access and munges the cumulative time series confirmed,
deaths and recovered from the data in the repository for the 2019 Novel Coronavirus Visual
Dashboard operated by the Johns Hopkins University Center for
Systems Science and Engineering (JHU CSSE). Also, Supported by ESRI
Living Atlas Team and the Johns Hopkins University Applied Physics
Lab (JHU APL).
}
\details{
Data are updated daily by JHU. Each call to this function redownloads the data
from github. No data cleansing is performed. Data are downloaded and then munged
into long-form tidy \code{data.frame}.
}
\note{
Uses https://raw.githubusercontent.com/CSSEGISandData/... as data
source, then modifies column names and munges to long form table.

\itemize{
\item US States are treated different from other countries, so are not directly included right now.
\item Although numbers are meant to be cumulative, there are instances where a day's count might
be less than the prior day due to a reclassification of a case. These are not currently corrected
in the source data
}
}
\examples{
res = jhu_data()
colnames(res)
head(res)
dplyr::glimpse(res)

}
\seealso{
Other data-import: 
\code{\link{apple_mobility_data}()},
\code{\link{beoutbreakprepared_data}()},
\code{\link{cdc_aggregated_projections}()},
\code{\link{cdc_excess_deaths}()},
\code{\link{cov_glue_lineage_data}()},
\code{\link{cov_glue_newick_data}()},
\code{\link{covidtracker_data}()},
\code{\link{descartes_mobility_data}()},
\code{\link{ecdc_data}()},
\code{\link{eu_data_cache_data}()},
\code{\link{google_mobility_data}()},
\code{\link{government_policy_timeline}()},
\code{\link{healthdata_projections_data}()},
\code{\link{jhu_us_data}()},
\code{\link{kff_icu_beds}()},
\code{\link{nytimes_county_data}()},
\code{\link{param_estimates_published}()},
\code{\link{us_county_geo_details}()},
\code{\link{us_healthcare_capacity}()},
\code{\link{us_state_distancing_policy}()},
\code{\link{usa_facts_data}()}

Other case-tracking: 
\code{\link{align_to_baseline}()},
\code{\link{beoutbreakprepared_data}()},
\code{\link{bulk_estimate_Rt}()},
\code{\link{combined_us_cases_data}()},
\code{\link{covidtracker_data}()},
\code{\link{ecdc_data}()},
\code{\link{estimate_Rt}()},
\code{\link{eu_data_cache_data}()},
\code{\link{nytimes_county_data}()},
\code{\link{plot_epicurve}()},
\code{\link{usa_facts_data}()}
}
\concept{case-tracking}
\concept{data-import}
