% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jhu_to_excel.R
\name{jhu_data_to_excel}
\alias{jhu_data_to_excel}
\title{Create +/- write excel format of enriched JHU global data}
\usage{
jhu_data_to_excel(
  dat = jhu_data(),
  file = NA,
  cols_to_remove = c("callingCodes", "altSpellings", "latlng", "demonym", "timezones",
    "nativeName", "currencies", "languages", "translations", "flag", "regionalBlocs"),
  ...
)
}
\arguments{
\item{dat}{a data.frame-like object with at least column
\code{CountryRegion} that will be joined with country data.}

\item{file}{character(1) filename to which to save excel
file. If specified, use the write.xlsx() function from the
openxlsx package to create (or overwrite) the file of that
name. The excel file will have one tab for each of the \code{subset}
records from the supplied data.frame in \code{dat}.}

\item{cols_to_remove}{a character vector of column names from
\code{\link[=country_metadata]{country_metadata()}} to remove.}
}
\value{
A list of three \code{data.frames} named \code{deaths}, \code{confirmed},
and \code{recovered}.
}
\description{
The purpose of this function is to enrich the JHU dataset
with additional country-level metadata and then get the
data into a form that can be easily written to excel or used
as a "wide-format" tabular dataset.
}
\examples{
xls_form = jhu_data_to_excel()
names(xls_form)
head(xls_form[[1]])

}
