% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owid_data.R
\name{owid_data}
\alias{owid_data}
\title{Our world in data worldwide cases and tests}
\source{
\itemize{
\item \url{https://ourworldindata.org/coronavirus}
}
}
\usage{
owid_data()
}
\description{
The OurWoldInData dataset includes country-level testing,
deaths, and confirmed cases for most countries in the world.
}
\examples{
res = owid_data()
colnames(res)

head(res)

dplyr::glimpse(res)
  
}
\references{
Max Roser, Hannah Ritchie, Esteban Ortiz-Ospina and Joe Hasell (2020) -
"Coronavirus Pandemic (COVID-19)". Published online at OurWorldInData.org.
Retrieved from: 'https://ourworldindata.org/coronavirus' \link{Online Resource}
}
\seealso{
Other case-tracking: 
\code{\link{align_to_baseline}()},
\code{\link{beoutbreakprepared_data}()},
\code{\link{bulk_estimate_Rt}()},
\code{\link{combined_us_cases_data}()},
\code{\link{covidtracker_data}()},
\code{\link{ecdc_data}()},
\code{\link{estimate_Rt}()},
\code{\link{eu_data_cache_data}()},
\code{\link{jhu_data}()},
\code{\link{nytimes_county_data}()},
\code{\link{plot_epicurve}()},
\code{\link{usa_facts_data}()}
}
\author{
Sean Davis \href{mailto:seandavi@gmail.com}{seandavi@gmail.com}
}
\concept{case-tracking}
