% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{plot_series}
\alias{plot_series}
\title{supersimple series plotter}
\usage{
plot_series(
  province = "",
  country,
  dataset = try(fetch_JHU_Data(as.data.frame = TRUE)),
  ...
)
}
\arguments{
\item{province}{character(1) must be found in dataset ProvinceState field, "" is typical for data aggregated only at country level, and that is the default}

\item{country}{character(1) must be found in CountryRegion field}

\item{dataset}{data.frame as returned by fetch_JHU_Data}

\item{\dots}{passed to base::plot}
}
\description{
supersimple series plotter
}
\note{
An effort is made to change dates used as column names to lubridate date objects
that are respected in plotting.
}
\examples{
plot_series(country="Italy")
}
