% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/us_counties_gazetteer.R
\name{us_county_geo_details}
\alias{us_county_geo_details}
\title{United States county-level geographic details}
\source{
\url{https://raw.githubusercontent.com/josh-byster/fips_lat_long/master/counties.txt}
}
\usage{
us_county_geo_details()
}
\value{
a data frame with county names, FIPS codes, areas, and lat/long
}
\description{
United States county-level geographic details
}
\examples{
usc = us_county_geo_details()
dplyr::glimpse(usc)
usc
summary(usc)


}
\seealso{
\url{https://www.census.gov/geographies/reference-files/time-series/geo/gazetteer-files.html}

Other data-import: 
\code{\link{apple_mobility_data}()},
\code{\link{beoutbreakprepared_data}()},
\code{\link{cdc_aggregated_projections}()},
\code{\link{covidtracker_data}()},
\code{\link{descartes_mobility_data}()},
\code{\link{ecdc_data}()},
\code{\link{eu_data_cache_data}()},
\code{\link{google_mobility_data}()},
\code{\link{government_policy_timeline}()},
\code{\link{healthdata_projections_data}()},
\code{\link{jhu_data}()},
\code{\link{jhu_us_data}()},
\code{\link{kff_icu_beds}()},
\code{\link{nytimes_county_data}()},
\code{\link{param_estimates_published}()},
\code{\link{us_healthcare_capacity}()},
\code{\link{us_state_distancing_policy}()},
\code{\link{usa_facts_data}()}
}
\concept{data-import}
