% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_care_map.R
\name{us_healthcare_capacity}
\alias{us_healthcare_capacity}
\title{United States healthcare system capacity by provider}
\source{
\itemize{
\item \url{https://github.com/covidcaremap/covid19-healthsystemcapacity}
\item \url{https://raw.githubusercontent.com/covidcaremap/covid19-healthsystemcapacity/master/data/published/us_healthcare_capacity-facility-CovidCareMap.csv}
}
}
\usage{
us_healthcare_capacity()
}
\description{
United States healthcare system capacity by provider
}
\details{
From the data providers:

Mapping existing and forecasted health system capacity gaps (beds,
staffing, ventilators, supplies) to care for surging numbers of
COVID19 patients (especially ICU-level care) at high spatiotemporal
resolution (by facility, daily, all USA to start).
}
\examples{
res = us_healthcare_capacity()
colnames(res)
dplyr::glimpse(res)

}
\seealso{
Other data-import: 
\code{\link{apple_mobility_data}()},
\code{\link{beoutbreakprepared_data}()},
\code{\link{cdc_aggregated_projections}()},
\code{\link{covidtracker_data}()},
\code{\link{descartes_mobility_data}()},
\code{\link{ecdc_data}()},
\code{\link{eu_data_cache_data}()},
\code{\link{google_mobility_data}()},
\code{\link{government_policy_timeline}()},
\code{\link{healthdata_projections_data}()},
\code{\link{jhu_data}()},
\code{\link{jhu_us_data}()},
\code{\link{kff_icu_beds}()},
\code{\link{nytimes_county_data}()},
\code{\link{param_estimates_published}()},
\code{\link{us_county_geo_details}()},
\code{\link{us_state_distancing_policy}()},
\code{\link{usa_facts_data}()}

Other healthcare-system: 
\code{\link{healthdata_projections_data}()}
}
\concept{data-import}
\concept{healthcare-system}
