% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/us_state_distancing_policy.R
\name{us_state_distancing_policy}
\alias{us_state_distancing_policy}
\title{United States social distancing policies for COVID-19}
\source{
\itemize{
\item https://github.com/COVID19StatePolicy/SocialDistancing/
}
}
\usage{
us_state_distancing_policy()
}
\description{
An updated, curated set of social distancing and non-pharmaceutical interventions
for states in the United States and District of Columbia.
}
\section{Kudos}{

\itemize{
\item Joe Wasserman: \url{http://www.joewasserman.com/}
}
}

\examples{
res = us_state_distancing_policy()
colnames(res)
library(dplyr)
dplyr::glimpse(res)
summary(res)

}
\seealso{
Other data-import: 
\code{\link{apple_mobility_data}()},
\code{\link{beoutbreakprepared_data}()},
\code{\link{cdc_aggregated_projections}()},
\code{\link{cdc_excess_deaths}()},
\code{\link{coronanet_government_response_data}()},
\code{\link{cov_glue_lineage_data}()},
\code{\link{cov_glue_newick_data}()},
\code{\link{covidtracker_data}()},
\code{\link{descartes_mobility_data}()},
\code{\link{ecdc_data}()},
\code{\link{eu_data_cache_data}()},
\code{\link{google_mobility_data}()},
\code{\link{government_policy_timeline}()},
\code{\link{healthdata_projections_data}()},
\code{\link{jhu_data}()},
\code{\link{jhu_us_data}()},
\code{\link{kff_icu_beds}()},
\code{\link{nytimes_county_data}()},
\code{\link{param_estimates_published}()},
\code{\link{us_county_geo_details}()},
\code{\link{us_county_health_rankings}()},
\code{\link{us_healthcare_capacity}()},
\code{\link{us_hospital_details}()},
\code{\link{usa_facts_data}()}
}
\author{
Sean Davis \href{mailto:seandavi@gmail.com}{seandavi@gmail.com}
}
\concept{data-import}
\concept{non-pharmaceutical-interventions}
